Split B+-tree module
