+++ src/core/btree.rs
@@ -4,7 +4,7 @@
 //! only with raw byte keys and raw byte values stored in `RawLeaf` pages
 //! and separator byte keys stored in `RawInterior` pages.
 
-use std::{cell::Cell, cmp::Ordering, fmt, ops::Range, rc::Rc};
+use std::{cell::Cell, cmp::Ordering, rc::Rc};
 
 use crate::core::{
     PAGE_SIZE, PageId,
@@ -23,277 +23,25 @@ use crate::core::{
     page_store::PageStore,
 };
 
-type MaterializedLeafCell = (Box<[u8]>, Box<[u8]>);
+mod mutation;
+mod payload;
+mod rebalance;
+mod record;
+mod root;
+mod search;
+mod split;
 
-/// Outcome of trying to position a scan within or beyond one leaf page.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum LeafSeek {
-    /// A concrete slot was found in the current leaf page.
-    Positioned(u16),
-    /// The scan must continue from an adjacent leaf page.
-    Advance(PageId),
-    /// No more leaf pages remain in the scan direction.
-    Exhausted,
-}
+#[cfg(test)]
+mod tests;
 
-/// Direction selector for shared cursor scans over linked leaf pages.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum ScanDirection {
-    /// Move toward larger keys.
-    Forward,
-    /// Move toward smaller keys.
-    Backward,
-}
+pub use record::{OwnedRecord, Record, RecordView};
+pub(crate) use root::{initialize_empty_root, validate_root_page};
 
-impl ScanDirection {
-    /// Descends from `start_page_id` to the leaf at the edge implied by `self`.
-    fn descend_to_edge_leaf<S: PageStore>(
-        self,
-        cursor: &TreeCursor<S>,
-        start_page_id: PageId,
-    ) -> StorageResult<PageId> {
-        match self {
-            Self::Forward => cursor.descend_to_first_leaf_from(start_page_id),
-            Self::Backward => cursor.descend_to_last_leaf_from(start_page_id),
-        }
-    }
+#[cfg(test)]
+use record::RecordStorage;
+#[cfg(test)]
+use root::read_page_kind;
 
-    /// Chooses the first or last slot in `leaf`, or advances to the next leaf
-    /// in the scan direction when the page is empty.
-    fn edge_seek(self, leaf: &RawLeaf<Read<'_>>) -> LeafSeek {
-        match self {
-            Self::Forward => {
-                if leaf.slot_count() > 0 {
-                    LeafSeek::Positioned(0)
-                } else if let Some(next_page_id) = leaf.next_page_id() {
-                    LeafSeek::Advance(next_page_id)
-                } else {
-                    LeafSeek::Exhausted
-                }
-            }
-            Self::Backward => {
-                if leaf.slot_count() > 0 {
-                    LeafSeek::Positioned(leaf.slot_count() - 1)
-                } else if let Some(prev_page_id) = leaf.prev_page_id() {
-                    LeafSeek::Advance(prev_page_id)
-                } else {
-                    LeafSeek::Exhausted
-                }
-            }
-        }
-    }
-
-    /// Chooses the adjacent slot relative to `slot_index`, or advances to the
-    /// neighboring leaf when the cursor is already at the page boundary.
-    fn adjacent_seek(self, leaf: &RawLeaf<Read<'_>>, slot_index: u16) -> LeafSeek {
-        match self {
-            Self::Forward => {
-                if slot_index + 1 < leaf.slot_count() {
-                    LeafSeek::Positioned(slot_index + 1)
-                } else if let Some(next_page_id) = leaf.next_page_id() {
-                    LeafSeek::Advance(next_page_id)
-                } else {
-                    LeafSeek::Exhausted
-                }
-            }
-            Self::Backward => {
-                if slot_index > 0 {
-                    LeafSeek::Positioned(slot_index - 1)
-                } else if let Some(prev_page_id) = leaf.prev_page_id() {
-                    LeafSeek::Advance(prev_page_id)
-                } else {
-                    LeafSeek::Exhausted
-                }
-            }
-        }
-    }
-}
-
-enum RecordStorage<S: PageStore> {
-    PageResident { pin: PinGuard<S>, key_range: Range<usize>, value_range: Range<usize> },
-    Materialized { key: Box<[u8]>, value: Box<[u8]> },
-}
-
-/// Raw record returned by tree reads and cursor iteration.
-///
-/// Records for cells that fit entirely in their leaf page borrow the page
-/// through an internal pin and only expose byte slices during accessor
-/// callbacks. Records for cells with overflow payload are materialized into
-/// fixed-size heap allocations. Use [`OwnedRecord`] when a stable snapshot is
-/// needed across later tree mutations.
-pub struct Record<S: PageStore> {
-    page_id: PageId,
-    slot_index: u16,
-    storage: RecordStorage<S>,
-}
-
-/// Stable, owned raw record snapshot.
-pub struct OwnedRecord {
-    key: Box<[u8]>,
-    value: Box<[u8]>,
-}
-
-/// Borrowed record view valid only for the callback that receives it.
-#[derive(Debug, Clone, Copy)]
-pub struct RecordView<'a> {
-    key: &'a [u8],
-    value: &'a [u8],
-}
-
-impl<'a> RecordView<'a> {
-    fn new(key: &'a [u8], value: &'a [u8]) -> Self {
-        Self { key, value }
-    }
-
-    /// Returns the record key bytes.
-    pub fn key(&self) -> &'a [u8] {
-        self.key
-    }
-
-    /// Returns the record value bytes.
-    pub fn value(&self) -> &'a [u8] {
-        self.value
-    }
-
-    /// Returns the record key and value bytes.
-    pub fn key_value(&self) -> (&'a [u8], &'a [u8]) {
-        (self.key, self.value)
-    }
-}
-
-impl<S: PageStore> Record<S> {
-    /// Builds a record view from one raw leaf-page slot.
-    pub(crate) fn new(
-        page_cache: &PageCache<S>,
-        page_id: PageId,
-        slot_index: u16,
-    ) -> StorageResult<Self> {
-        let pin = page_cache.fetch_page(page_id)?;
-        let (key_len, value_len, first_overflow_page_id, inline_range) = {
-            let page = pin.read()?;
-            let leaf = page.open::<Leaf>()?;
-            leaf.cell_payload_parts(slot_index)?
-        };
-
-        let storage = match first_overflow_page_id {
-            None => {
-                if inline_range.len() != key_len + value_len {
-                    return Err(cell_corruption(page_id, CorruptionKind::CellLengthOutOfBounds));
-                }
-
-                let key_start = inline_range.start;
-                let value_start = key_start + key_len;
-                RecordStorage::PageResident {
-                    pin,
-                    key_range: key_start..value_start,
-                    value_range: value_start..inline_range.end,
-                }
-            }
-            Some(first_overflow_page_id) => {
-                let (key, value) = {
-                    let page = pin.read()?;
-                    materialize_leaf_cell(
-                        page_cache,
-                        page_id,
-                        &page.page()[inline_range],
-                        first_overflow_page_id,
-                        key_len,
-                        value_len,
-                    )?
-                };
-                drop(pin);
-
-                RecordStorage::Materialized { key, value }
-            }
-        };
-
-        Ok(Self { page_id, slot_index, storage })
-    }
-
-    /// Returns the slot index that this record refers to within its leaf page.
-    pub fn slot_index(&self) -> u16 {
-        self.slot_index
-    }
-
-    /// Executes `f` with a borrowed view of this record.
-    pub fn with_view<R>(&self, f: impl FnOnce(RecordView<'_>) -> R) -> StorageResult<R> {
-        match &self.storage {
-            RecordStorage::PageResident { pin, key_range, value_range } => {
-                let page = pin.read()?;
-                let key = &page.page()[key_range.clone()];
-                let value = &page.page()[value_range.clone()];
-                Ok(f(RecordView::new(key, value)))
-            }
-            RecordStorage::Materialized { key, value } => {
-                Ok(f(RecordView::new(key.as_ref(), value.as_ref())))
-            }
-        }
-    }
-
-    /// Returns a stable, owned snapshot of this record.
-    pub fn to_owned_record(&self) -> StorageResult<OwnedRecord> {
-        self.with_key_value(|key, value| OwnedRecord::new(key.into(), value.into()))
-    }
-
-    /// Executes `f` with a borrowed view of the record key.
-    pub fn with_key<R>(&self, f: impl FnOnce(&[u8]) -> R) -> StorageResult<R> {
-        self.with_view(|record| f(record.key()))
-    }
-
-    /// Executes `f` with a borrowed view of the record value.
-    pub fn with_value<R>(&self, f: impl FnOnce(&[u8]) -> R) -> StorageResult<R> {
-        self.with_view(|record| f(record.value()))
-    }
-
-    /// Executes `f` with borrowed views of the key and value.
-    pub fn with_key_value<R>(&self, f: impl FnOnce(&[u8], &[u8]) -> R) -> StorageResult<R> {
-        self.with_view(|record| {
-            let (key, value) = record.key_value();
-            f(key, value)
-        })
-    }
-}
-
-impl OwnedRecord {
-    fn new(key: Box<[u8]>, value: Box<[u8]>) -> Self {
-        Self { key, value }
-    }
-
-    /// Executes `f` with a borrowed view of the record key.
-    pub fn with_key<R>(&self, f: impl FnOnce(&[u8]) -> R) -> R {
-        f(&self.key)
-    }
-
-    /// Executes `f` with a borrowed view of the record value.
-    pub fn with_value<R>(&self, f: impl FnOnce(&[u8]) -> R) -> R {
-        f(&self.value)
-    }
-
-    /// Executes `f` with borrowed views of the key and value.
-    pub fn with_key_value<R>(&self, f: impl FnOnce(&[u8], &[u8]) -> R) -> R {
-        f(&self.key, &self.value)
-    }
-}
-
-impl fmt::Debug for OwnedRecord {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("OwnedRecord")
-            .field("key_len", &self.key.len())
-            .field("value_len", &self.value.len())
-            .finish()
-    }
-}
-
-impl<S: PageStore> fmt::Debug for Record<S> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("Record")
-            .field("page_id", &self.page_id)
-            .field("slot_index", &self.slot_index)
-            .finish_non_exhaustive()
-    }
-}
-
-/// Logical cursor state exposed by the public cursor API.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum CursorState {
     /// The cursor is anchored to a page but not yet to a specific slot.
@@ -396,3054 +144,4 @@ fn payload_uses_overflow(payload_len: usize) -> bool {
 
 fn local_payload_len(payload_len: usize) -> usize {
     if payload_uses_overflow(payload_len) { MAX_INLINE_OVERFLOW_PAYLOAD_BYTES } else { payload_len }
-}
-
-fn cell_corruption(page_id: PageId, kind: CorruptionKind) -> StorageError {
-    StorageError::Corruption(CorruptionError {
-        component: CorruptionComponent::Cell,
-        page_id: Some(page_id),
-        kind,
-    })
-}
-
-fn overflow_corruption(page_id: Option<PageId>, kind: CorruptionKind) -> StorageError {
-    StorageError::Corruption(CorruptionError {
-        component: CorruptionComponent::OverflowPage,
-        page_id,
-        kind,
-    })
-}
-
-fn read_overflow_next_page_id(page: &[u8; PAGE_SIZE]) -> Option<PageId> {
-    page::format::read_optional_u64(page, 0)
-}
-
-fn write_overflow_chain_from_slices<S: PageStore>(
-    page_cache: &PageCache<S>,
-    mut first: &[u8],
-    mut second: &[u8],
-) -> StorageResult<Option<PageId>> {
-    if first.is_empty() && second.is_empty() {
-        return Ok(None);
-    }
-
-    let mut first_page_id = None;
-    let mut previous_page_id = None;
-    while !first.is_empty() || !second.is_empty() {
-        let (page_id, pin) = page_cache.new_page()?;
-        {
-            let mut page = pin.write()?;
-            page.page_mut().fill(0);
-            page::format::write_optional_u64(page.page_mut(), 0, None);
-
-            let mut write_offset = OVERFLOW_NEXT_PAGE_ID_SIZE;
-            while write_offset < PAGE_SIZE && (!first.is_empty() || !second.is_empty()) {
-                let source = if !first.is_empty() { first } else { second };
-                let take = source.len().min(PAGE_SIZE - write_offset);
-                page.page_mut()[write_offset..write_offset + take].copy_from_slice(&source[..take]);
-                write_offset += take;
-                if !first.is_empty() {
-                    first = &first[take..];
-                } else {
-                    second = &second[take..];
-                }
-            }
-        }
-        drop(pin);
-
-        if first_page_id.is_none() {
-            first_page_id = Some(page_id);
-        }
-        if let Some(previous_page_id) = previous_page_id {
-            let previous_pin = page_cache.fetch_page(previous_page_id)?;
-            let mut previous_page = previous_pin.write()?;
-            page::format::write_optional_u64(previous_page.page_mut(), 0, Some(page_id));
-        }
-        previous_page_id = Some(page_id);
-    }
-
-    Ok(first_page_id)
-}
-
-fn append_overflow_chain_exact<S: PageStore>(
-    page_cache: &PageCache<S>,
-    first_overflow_page_id: PageId,
-    payload: &mut Vec<u8>,
-    total_payload_len: usize,
-) -> StorageResult<()> {
-    let initial_len = payload.len();
-    let expected_overflow_len = total_payload_len - initial_len;
-    let mut page_id = Some(first_overflow_page_id);
-
-    while payload.len() < total_payload_len {
-        let Some(current_page_id) = page_id else {
-            return Err(overflow_corruption(
-                None,
-                CorruptionKind::OverflowChainTooShort {
-                    expected: expected_overflow_len,
-                    actual: payload.len() - initial_len,
-                },
-            ));
-        };
-
-        if current_page_id == NO_OVERFLOW_PAGE_ID {
-            return Err(overflow_corruption(
-                Some(current_page_id),
-                CorruptionKind::OverflowChainTooShort {
-                    expected: expected_overflow_len,
-                    actual: payload.len() - initial_len,
-                },
-            ));
-        }
-
-        let pin = page_cache.fetch_page(current_page_id)?;
-        let page = pin.read()?;
-        let remaining = total_payload_len - payload.len();
-        let take = remaining.min(overflow::OVERFLOW_PAYLOAD_SIZE);
-        payload.extend_from_slice(
-            &page.page()[OVERFLOW_NEXT_PAGE_ID_SIZE..OVERFLOW_NEXT_PAGE_ID_SIZE + take],
-        );
-        page_id = read_overflow_next_page_id(page.page());
-    }
-
-    if page_id.is_some() {
-        return Err(overflow_corruption(
-            page_id,
-            CorruptionKind::OverflowChainTooLong { expected: expected_overflow_len },
-        ));
-    }
-
-    Ok(())
-}
-
-fn materialize_payload<S: PageStore>(
-    page_cache: &PageCache<S>,
-    page_id: PageId,
-    inline_payload: &[u8],
-    first_overflow_page_id: Option<PageId>,
-    payload_len: usize,
-) -> StorageResult<Vec<u8>> {
-    if inline_payload.len() > payload_len {
-        return Err(cell_corruption(page_id, CorruptionKind::CellLengthOutOfBounds));
-    }
-
-    let mut payload = Vec::with_capacity(payload_len);
-    payload.extend_from_slice(inline_payload);
-    match first_overflow_page_id {
-        Some(first_overflow_page_id) => append_overflow_chain_exact(
-            page_cache,
-            first_overflow_page_id,
-            &mut payload,
-            payload_len,
-        )?,
-        None if payload.len() != payload_len => {
-            return Err(cell_corruption(page_id, CorruptionKind::CellLengthOutOfBounds));
-        }
-        None => {}
-    }
-
-    if payload.len() != payload_len {
-        return Err(cell_corruption(page_id, CorruptionKind::CellLengthOutOfBounds));
-    }
-    Ok(payload)
-}
-
-fn materialize_leaf_cell<S: PageStore>(
-    page_cache: &PageCache<S>,
-    page_id: PageId,
-    inline_payload: &[u8],
-    first_overflow_page_id: PageId,
-    key_len: usize,
-    value_len: usize,
-) -> StorageResult<MaterializedLeafCell> {
-    let mut payload = materialize_payload(
-        page_cache,
-        page_id,
-        inline_payload,
-        Some(first_overflow_page_id),
-        key_len + value_len,
-    )?;
-    if payload.len() < key_len {
-        return Err(cell_corruption(page_id, CorruptionKind::CellLengthOutOfBounds));
-    }
-
-    let value = payload.split_off(key_len);
-    Ok((payload.into_boxed_slice(), value.into_boxed_slice()))
-}
-
-fn read_leaf_cell<S: PageStore>(
-    page_cache: &PageCache<S>,
-    page_id: PageId,
-    slot_index: u16,
-) -> StorageResult<(Vec<u8>, Vec<u8>)> {
-    let pin = page_cache.fetch_page(page_id)?;
-    let payload = {
-        let page = pin.read()?;
-        let leaf = page.open::<Leaf>()?;
-        let (key_len, value_len, first_overflow_page_id, inline_range) =
-            leaf.cell_payload_parts(slot_index)?;
-        let mut payload = materialize_payload(
-            page_cache,
-            page_id,
-            &page.page()[inline_range],
-            first_overflow_page_id,
-            key_len + value_len,
-        )?;
-        if payload.len() < key_len {
-            return Err(cell_corruption(page_id, CorruptionKind::CellLengthOutOfBounds));
-        }
-        let value = payload.split_off(key_len);
-        (payload, value)
-    };
-    drop(pin);
-    Ok(payload)
-}
-
-fn read_interior_cell<S: PageStore>(
-    page_cache: &PageCache<S>,
-    page_id: PageId,
-    slot_index: u16,
-) -> StorageResult<(PageId, Vec<u8>)> {
-    let pin = page_cache.fetch_page(page_id)?;
-    let (left_child, key) = {
-        let page = pin.read()?;
-        let interior = page.open::<Interior>()?;
-        let (left_child, key_len, first_overflow_page_id, inline_range) =
-            interior.cell_payload_parts(slot_index)?;
-        let key = materialize_payload(
-            page_cache,
-            page_id,
-            &page.page()[inline_range],
-            first_overflow_page_id,
-            key_len,
-        )?;
-        (left_child, key)
-    };
-    drop(pin);
-
-    Ok((left_child, key))
-}
-
-fn compare_key_prefix(
-    page_id: PageId,
-    inline_key: &[u8],
-    key_len: usize,
-    key: &[u8],
-) -> StorageResult<Option<Ordering>> {
-    if inline_key.len() > key_len {
-        return Err(cell_corruption(page_id, CorruptionKind::CellLengthOutOfBounds));
-    }
-
-    if key.len() <= inline_key.len() {
-        let ordering = inline_key[..key.len()].cmp(key);
-        if ordering != Ordering::Equal {
-            return Ok(Some(ordering));
-        }
-        return Ok(Some(key_len.cmp(&key.len())));
-    }
-
-    let ordering = inline_key.cmp(&key[..inline_key.len()]);
-    if ordering != Ordering::Equal {
-        return Ok(Some(ordering));
-    }
-
-    if inline_key.len() == key_len {
-        return Ok(Some(Ordering::Less));
-    }
-
-    Ok(None)
-}
-
-fn compare_overflow_key<S: PageStore>(
-    page_cache: &PageCache<S>,
-    page_id: PageId,
-    key: &[u8],
-    inline_key_len: usize,
-    first_overflow_page_id: Option<PageId>,
-    key_len: usize,
-) -> StorageResult<Ordering> {
-    if inline_key_len > key_len {
-        return Err(cell_corruption(page_id, CorruptionKind::CellLengthOutOfBounds));
-    }
-
-    let first_overflow_page_id = first_overflow_page_id
-        .ok_or_else(|| cell_corruption(page_id, CorruptionKind::CellLengthOutOfBounds))?;
-
-    let mut page_id = Some(first_overflow_page_id);
-    let mut compared = inline_key_len;
-    while compared < key_len {
-        let Some(current_page_id) = page_id else {
-            return Err(overflow_corruption(
-                None,
-                CorruptionKind::OverflowChainTooShort {
-                    expected: key_len - inline_key_len,
-                    actual: compared - inline_key_len,
-                },
-            ));
-        };
-
-        if current_page_id == NO_OVERFLOW_PAGE_ID {
-            return Err(overflow_corruption(
-                Some(current_page_id),
-                CorruptionKind::OverflowChainTooShort {
-                    expected: key_len - inline_key_len,
-                    actual: compared - inline_key_len,
-                },
-            ));
-        }
-
-        let pin = page_cache.fetch_page(current_page_id)?;
-        let page = pin.read()?;
-        let remaining = key_len - compared;
-        let take = remaining.min(overflow::OVERFLOW_PAYLOAD_SIZE);
-        if compared < key.len() {
-            let compare_len = (key.len() - compared).min(take);
-            let ordering = page.page()
-                [OVERFLOW_NEXT_PAGE_ID_SIZE..OVERFLOW_NEXT_PAGE_ID_SIZE + compare_len]
-                .cmp(&key[compared..compared + compare_len]);
-            if ordering != Ordering::Equal {
-                return Ok(ordering);
-            }
-        }
-
-        compared += take;
-        page_id = read_overflow_next_page_id(page.page());
-    }
-
-    Ok(key_len.cmp(&key.len()))
-}
-
-impl<S: PageStore> TreeCursor<S> {
-    /// Creates a cursor anchored at `root_page_id` in page-level state.
-    pub(crate) fn new(page_cache: PageCache<S>, root_page_id: PageId) -> Self {
-        Self {
-            page_cache,
-            root_page_id: Rc::new(Cell::new(root_page_id)),
-            state: CursorState::Page { page_id: root_page_id },
-        }
-    }
-
-    /// Returns the root page id that anchors this tree.
-    pub fn root_page_id(&self) -> PageId {
-        self.root_page_id.get()
-    }
-
-    /// Returns the cursor's current logical state.
-    pub fn state(&self) -> CursorState {
-        self.state
-    }
-
-    /// Returns the page currently referenced by the cursor, if any.
-    pub fn current_page_id(&self) -> Option<PageId> {
-        match self.state {
-            CursorState::Page { page_id } | CursorState::Positioned { page_id, .. } => {
-                Some(page_id)
-            }
-            CursorState::Exhausted => None,
-        }
-    }
-
-    /// Returns `true` when the cursor is currently positioned on a record.
-    pub fn is_positioned(&self) -> bool {
-        matches!(self.state, CursorState::Positioned { .. })
-    }
-
-    /// Resets the cursor back to the tree root page.
-    pub fn seek_to_root(&mut self) {
-        self.state = CursorState::Page { page_id: self.root_page_id() };
-    }
-
-    /// Switches the cursor to a page-anchored but not slot-anchored state.
-    fn set_page_state(&mut self, page_id: PageId) {
-        self.state = CursorState::Page { page_id };
-    }
-
-    /// Switches the cursor to one concrete slot inside a leaf page.
-    fn set_positioned_state(&mut self, page_id: PageId, slot_index: u16) {
-        self.state = CursorState::Positioned { page_id, slot_index };
-    }
-
-    /// Marks the cursor as having moved past the scan range.
-    fn set_exhausted_state(&mut self) {
-        self.state = CursorState::Exhausted;
-    }
-
-    /// Materializes one record from a positioned raw leaf slot.
-    fn record_at(&self, page_id: PageId, slot_index: u16) -> StorageResult<Record<S>> {
-        Record::new(&self.page_cache, page_id, slot_index)
-    }
-
-    fn raw_leaf_slot_count(&self, page_id: PageId) -> StorageResult<u16> {
-        let pin = self.page_cache.fetch_page(page_id)?;
-        let page = pin.read()?;
-        let leaf = page.open::<Leaf>()?;
-        Ok(leaf.slot_count())
-    }
-
-    fn raw_interior_slot_count(&self, page_id: PageId) -> StorageResult<u16> {
-        let pin = self.page_cache.fetch_page(page_id)?;
-        let page = pin.read()?;
-        let interior = page.open::<Interior>()?;
-        Ok(interior.slot_count())
-    }
-
-    fn compare_leaf_key(
-        &self,
-        page_id: PageId,
-        slot_index: u16,
-        key: &[u8],
-    ) -> StorageResult<Ordering> {
-        let (inline_key_len, first_overflow_page_id, key_len) = {
-            let pin = self.page_cache.fetch_page(page_id)?;
-            let page = pin.read()?;
-            let leaf = page.open::<Leaf>()?;
-            let (key_len, _, first_overflow_page_id, inline_range) =
-                leaf.cell_payload_parts(slot_index)?;
-            let inline_key_len = key_len.min(inline_range.len());
-            let inline_key = &page.page()[inline_range.start..inline_range.start + inline_key_len];
-            if let Some(ordering) = compare_key_prefix(page_id, inline_key, key_len, key)? {
-                return Ok(ordering);
-            }
-            (inline_key_len, first_overflow_page_id, key_len)
-        };
-
-        compare_overflow_key(
-            &self.page_cache,
-            page_id,
-            key,
-            inline_key_len,
-            first_overflow_page_id,
-            key_len,
-        )
-    }
-
-    fn compare_interior_key(
-        &self,
-        page_id: PageId,
-        slot_index: u16,
-        key: &[u8],
-    ) -> StorageResult<Ordering> {
-        let (inline_key_len, first_overflow_page_id, key_len) = {
-            let pin = self.page_cache.fetch_page(page_id)?;
-            let page = pin.read()?;
-            let interior = page.open::<Interior>()?;
-            let (_, key_len, first_overflow_page_id, inline_range) =
-                interior.cell_payload_parts(slot_index)?;
-            let inline_key_len = key_len.min(inline_range.len());
-            let inline_key = &page.page()[inline_range.start..inline_range.start + inline_key_len];
-            if let Some(ordering) = compare_key_prefix(page_id, inline_key, key_len, key)? {
-                return Ok(ordering);
-            }
-            (inline_key_len, first_overflow_page_id, key_len)
-        };
-
-        compare_overflow_key(
-            &self.page_cache,
-            page_id,
-            key,
-            inline_key_len,
-            first_overflow_page_id,
-            key_len,
-        )
-    }
-
-    fn read_interior_left_child(&self, page_id: PageId, slot_index: u16) -> StorageResult<PageId> {
-        let pin = self.page_cache.fetch_page(page_id)?;
-        let page = pin.read()?;
-        let interior = page.open::<Interior>()?;
-        let (left_child, _, _, _) = interior.cell_payload_parts(slot_index)?;
-        Ok(left_child)
-    }
-
-    fn search_leaf_slot(&self, page_id: PageId, key: &[u8]) -> StorageResult<SearchResult> {
-        let mut low: u16 = 0;
-        let mut high = self.raw_leaf_slot_count(page_id)?;
-
-        while low < high {
-            let mid = low + (high - low) / 2;
-            match self.compare_leaf_key(page_id, mid, key)? {
-                Ordering::Less => low = mid + 1,
-                Ordering::Greater => high = mid,
-                Ordering::Equal => return Ok(SearchResult::Found(mid)),
-            }
-        }
-
-        Ok(SearchResult::InsertAt(low))
-    }
-
-    fn lower_bound_leaf_slot(&self, page_id: PageId, key: &[u8]) -> StorageResult<BoundResult> {
-        let mut low: u16 = 0;
-        let slot_count = self.raw_leaf_slot_count(page_id)?;
-        let mut high = slot_count;
-
-        while low < high {
-            let mid = low + (high - low) / 2;
-            if self.compare_leaf_key(page_id, mid, key)? == Ordering::Less {
-                low = mid + 1;
-            } else {
-                high = mid;
-            }
-        }
-
-        if low == slot_count { Ok(BoundResult::PastEnd) } else { Ok(BoundResult::At(low)) }
-    }
-
-    fn lower_bound_interior_slot(&self, page_id: PageId, key: &[u8]) -> StorageResult<BoundResult> {
-        let mut low: u16 = 0;
-        let slot_count = self.raw_interior_slot_count(page_id)?;
-        let mut high = slot_count;
-
-        while low < high {
-            let mid = low + (high - low) / 2;
-            if self.compare_interior_key(page_id, mid, key)? == Ordering::Less {
-                low = mid + 1;
-            } else {
-                high = mid;
-            }
-        }
-
-        if low == slot_count { Ok(BoundResult::PastEnd) } else { Ok(BoundResult::At(low)) }
-    }
-
-    fn interior_child_for_key(&self, page_id: PageId, key: &[u8]) -> StorageResult<PageId> {
-        match self.lower_bound_interior_slot(page_id, key)? {
-            BoundResult::At(slot_index) => self.read_interior_left_child(page_id, slot_index),
-            BoundResult::PastEnd => {
-                let pin = self.page_cache.fetch_page(page_id)?;
-                let page = pin.read()?;
-                let interior = page.open::<Interior>()?;
-                Ok(interior.rightmost_child())
-            }
-        }
-    }
-
-    /// Follows leftmost children from `start_page_id` until reaching a leaf.
-    fn descend_to_first_leaf_from(&self, start_page_id: PageId) -> StorageResult<PageId> {
-        let mut page_id = start_page_id;
-
-        loop {
-            let pin = self.page_cache.fetch_page(page_id)?;
-            let next = {
-                let page = pin.read()?;
-                match read_page_kind(page.page(), page_id)? {
-                    PageKind::RawLeaf => {
-                        let _ = page.open::<Leaf>()?;
-                        return Ok(page_id);
-                    }
-                    PageKind::RawInterior => {
-                        let interior = page.open::<Interior>()?;
-                        if interior.slot_count() == 0 {
-                            interior.rightmost_child()
-                        } else {
-                            interior.cell(0)?.left_child()?
-                        }
-                    }
-                }
-            };
-            page_id = next;
-        }
-    }
-
-    /// Follows rightmost children from `start_page_id` until reaching a leaf.
-    fn descend_to_last_leaf_from(&self, start_page_id: PageId) -> StorageResult<PageId> {
-        let mut page_id = start_page_id;
-
-        loop {
-            let pin = self.page_cache.fetch_page(page_id)?;
-            let next = {
-                let page = pin.read()?;
-                match read_page_kind(page.page(), page_id)? {
-                    PageKind::RawLeaf => {
-                        let _ = page.open::<Leaf>()?;
-                        return Ok(page_id);
-                    }
-                    PageKind::RawInterior => {
-                        let interior = page.open::<Interior>()?;
-                        interior.rightmost_child()
-                    }
-                }
-            };
-            page_id = next;
-        }
-    }
-
-    /// Descends to the leaf page that contains or would contain `key`.
-    fn leaf_page_for_key(&self, key: &[u8]) -> StorageResult<PageId> {
-        let mut page_id = self.root_page_id();
-
-        loop {
-            let pin = self.page_cache.fetch_page(page_id)?;
-            let next = {
-                let page = pin.read()?;
-                match read_page_kind(page.page(), page_id)? {
-                    PageKind::RawLeaf => {
-                        let _ = page.open::<Leaf>()?;
-                        return Ok(page_id);
-                    }
-                    PageKind::RawInterior => {
-                        drop(page);
-                        drop(pin);
-                        self.interior_child_for_key(page_id, key)?
-                    }
-                }
-            };
-            page_id = next;
-        }
-    }
-
-    /// Descends to the target leaf and records the interior path taken to reach it.
-    fn leaf_page_path_for_key(&self, key: &[u8]) -> StorageResult<(PageId, Vec<PathFrame>)> {
-        let mut path = Vec::new();
-        let mut page_id = self.root_page_id();
-
-        loop {
-            let pin = self.page_cache.fetch_page(page_id)?;
-            let next = {
-                let page = pin.read()?;
-                match read_page_kind(page.page(), page_id)? {
-                    PageKind::RawLeaf => {
-                        let _ = page.open::<Leaf>()?;
-                        return Ok((page_id, path));
-                    }
-                    PageKind::RawInterior => {
-                        drop(page);
-                        drop(pin);
-                        match self.lower_bound_interior_slot(page_id, key)? {
-                            BoundResult::At(slot_index) => {
-                                let child_page_id =
-                                    self.read_interior_left_child(page_id, slot_index)?;
-                                path.push(PathFrame {
-                                    page_id,
-                                    child_ref: ChildSlotRef::Slot(slot_index),
-                                });
-                                child_page_id
-                            }
-                            BoundResult::PastEnd => {
-                                path.push(PathFrame {
-                                    page_id,
-                                    child_ref: ChildSlotRef::Rightmost,
-                                });
-                                let pin = self.page_cache.fetch_page(page_id)?;
-                                let page = pin.read()?;
-                                let interior = page.open::<Interior>()?;
-                                interior.rightmost_child()
-                            }
-                        }
-                    }
-                }
-            };
-            page_id = next;
-        }
-    }
-
-    /// Reads the first reachable record from `start_page_id` in `direction`,
-    /// skipping over empty leaf pages until a slot is found or the scan ends.
-    fn edge_record_from_leaf(
-        &mut self,
-        start_page_id: PageId,
-        direction: ScanDirection,
-    ) -> StorageResult<Option<Record<S>>> {
-        let mut page_id = start_page_id;
-
-        loop {
-            let pin = self.page_cache.fetch_page(page_id)?;
-            let seek = {
-                let page = pin.read()?;
-                expect_page_kind(page.page(), page_id, PageKind::RawLeaf, "raw leaf")?;
-                let leaf = page.open::<Leaf>()?;
-                direction.edge_seek(&leaf)
-            };
-
-            match seek {
-                LeafSeek::Positioned(slot_index) => {
-                    self.set_positioned_state(page_id, slot_index);
-                    return self.record_at(page_id, slot_index).map(Some);
-                }
-                LeafSeek::Advance(next_page_id) => page_id = next_page_id,
-                LeafSeek::Exhausted => {
-                    self.set_exhausted_state();
-                    return Ok(None);
-                }
-            }
-        }
-    }
-
-    /// Advances or rewinds the cursor by one logical record.
-    fn step_record(&mut self, direction: ScanDirection) -> StorageResult<Option<Record<S>>> {
-        match self.state {
-            CursorState::Exhausted => Ok(None),
-            CursorState::Page { page_id } => {
-                let leaf_page_id = direction.descend_to_edge_leaf(self, page_id)?;
-                self.edge_record_from_leaf(leaf_page_id, direction)
-            }
-            CursorState::Positioned { page_id, slot_index } => {
-                let pin = self.page_cache.fetch_page(page_id)?;
-                let seek = {
-                    let page = pin.read()?;
-                    expect_page_kind(page.page(), page_id, PageKind::RawLeaf, "raw leaf")?;
-                    let leaf = page.open::<Leaf>()?;
-                    direction.adjacent_seek(&leaf, slot_index)
-                };
-
-                match seek {
-                    LeafSeek::Positioned(next_slot) => {
-                        self.set_positioned_state(page_id, next_slot);
-                        self.record_at(page_id, next_slot).map(Some)
-                    }
-                    LeafSeek::Advance(next_page_id) => {
-                        self.edge_record_from_leaf(next_page_id, direction)
-                    }
-                    LeafSeek::Exhausted => {
-                        self.set_exhausted_state();
-                        Ok(None)
-                    }
-                }
-            }
-        }
-    }
-
-    /// Searches the raw tree for `key`.
-    ///
-    /// The cursor ends on the matching record when found, or on the leaf page
-    /// where `key` would be inserted when absent.
-    pub fn get(&mut self, key: &[u8]) -> StorageResult<Option<Record<S>>> {
-        let page_id = self.leaf_page_for_key(key)?;
-        let slot_index = match self.search_leaf_slot(page_id, key)? {
-            SearchResult::Found(slot_index) => Some(slot_index),
-            SearchResult::InsertAt(_) => None,
-        };
-
-        match slot_index {
-            Some(slot_index) => {
-                self.set_positioned_state(page_id, slot_index);
-                self.record_at(page_id, slot_index).map(Some)
-            }
-            None => {
-                self.set_page_state(page_id);
-                Ok(None)
-            }
-        }
-    }
-
-    /// Searches the raw tree for `key` and returns a stable owned record snapshot.
-    pub fn get_owned(&mut self, key: &[u8]) -> StorageResult<Option<OwnedRecord>> {
-        self.get(key)?.map(|record| record.to_owned_record()).transpose()
-    }
-
-    fn checked_payload_len(&self, payload_len: usize) -> StorageResult<()> {
-        if payload_len > u16::MAX as usize {
-            return Err(PageError::CellTooLarge { len: payload_len, max: u16::MAX as usize }.into());
-        }
-        Ok(())
-    }
-
-    fn leaf_cell_local_size(&self, key: &[u8], value: &[u8]) -> StorageResult<usize> {
-        let payload_len = key.len() + value.len();
-        self.checked_payload_len(payload_len)?;
-        Ok(LEAF_CELL_PREFIX_SIZE + local_payload_len(payload_len))
-    }
-
-    fn interior_cell_local_size(&self, key: &[u8]) -> StorageResult<usize> {
-        self.checked_payload_len(key.len())?;
-        Ok(INTERIOR_CELL_PREFIX_SIZE + local_payload_len(key.len()))
-    }
-
-    /// Builds the local leaf payload bytes and optional overflow chain for one key/value cell.
-    fn prepare_leaf_payload(
-        &self,
-        key: &[u8],
-        value: &[u8],
-    ) -> StorageResult<(Option<PageId>, Vec<u8>)> {
-        let payload_len = key.len() + value.len();
-        self.checked_payload_len(payload_len)?;
-        let mut inline_payload = [0; MAX_INLINE_OVERFLOW_PAYLOAD_BYTES];
-
-        let first_overflow_page_id = if payload_uses_overflow(payload_len) {
-            if key.len() >= MAX_INLINE_OVERFLOW_PAYLOAD_BYTES {
-                inline_payload.copy_from_slice(&key[..MAX_INLINE_OVERFLOW_PAYLOAD_BYTES]);
-                Some(
-                    write_overflow_chain_from_slices(
-                        &self.page_cache,
-                        &key[MAX_INLINE_OVERFLOW_PAYLOAD_BYTES..],
-                        value,
-                    )?
-                    .ok_or_else(|| {
-                        cell_corruption(self.root_page_id(), CorruptionKind::CellLengthOutOfBounds)
-                    })?,
-                )
-            } else {
-                inline_payload[..key.len()].copy_from_slice(key);
-                let value_prefix_len = MAX_INLINE_OVERFLOW_PAYLOAD_BYTES - key.len();
-                inline_payload[key.len()..MAX_INLINE_OVERFLOW_PAYLOAD_BYTES]
-                    .copy_from_slice(&value[..value_prefix_len]);
-                Some(
-                    overflow::write_chain(&self.page_cache, &value[value_prefix_len..])?
-                        .ok_or_else(|| {
-                            cell_corruption(
-                                self.root_page_id(),
-                                CorruptionKind::CellLengthOutOfBounds,
-                            )
-                        })?,
-                )
-            }
-        } else {
-            inline_payload[..key.len()].copy_from_slice(key);
-            inline_payload[key.len()..payload_len].copy_from_slice(value);
-            None
-        };
-        let inline_payload = inline_payload[..local_payload_len(payload_len)].to_vec();
-        Ok((first_overflow_page_id, inline_payload))
-    }
-
-    fn insert_leaf_payload_at(
-        &self,
-        leaf: &mut RawLeaf<Write<'_>>,
-        slot_index: u16,
-        key: &[u8],
-        value: &[u8],
-    ) -> StorageResult<u16> {
-        let (first_overflow_page_id, inline_payload) = self.prepare_leaf_payload(key, value)?;
-        Ok(leaf.insert_payload_at(
-            slot_index,
-            key.len(),
-            value.len(),
-            first_overflow_page_id,
-            &inline_payload,
-        )?)
-    }
-
-    /// Rewrites one existing leaf slot using the same overflow layout as tree inserts.
-    fn update_leaf_payload_at(
-        &self,
-        leaf: &mut RawLeaf<Write<'_>>,
-        slot_index: u16,
-        key: &[u8],
-        value: &[u8],
-    ) -> StorageResult<u16> {
-        let (first_overflow_page_id, inline_payload) = self.prepare_leaf_payload(key, value)?;
-        Ok(leaf.update_payload_at(
-            slot_index,
-            key.len(),
-            value.len(),
-            first_overflow_page_id,
-            &inline_payload,
-        )?)
-    }
-
-    fn insert_interior_payload_at(
-        &self,
-        interior: &mut RawInterior<Write<'_>>,
-        slot_index: u16,
-        left_child: PageId,
-        key: &[u8],
-    ) -> StorageResult<u16> {
-        self.checked_payload_len(key.len())?;
-        let (first_overflow_page_id, inline_payload): (Option<PageId>, &[u8]) =
-            if payload_uses_overflow(key.len()) {
-                (
-                    Some(
-                        overflow::write_chain(
-                            &self.page_cache,
-                            &key[MAX_INLINE_OVERFLOW_PAYLOAD_BYTES..],
-                        )?
-                        .ok_or_else(|| {
-                            cell_corruption(
-                                self.root_page_id(),
-                                CorruptionKind::CellLengthOutOfBounds,
-                            )
-                        })?,
-                    ),
-                    &key[..MAX_INLINE_OVERFLOW_PAYLOAD_BYTES],
-                )
-            } else {
-                (None, key)
-            };
-        Ok(interior.insert_payload_at(
-            slot_index,
-            left_child,
-            key.len(),
-            first_overflow_page_id,
-            inline_payload,
-        )?)
-    }
-
-    fn missing_child_max_key_error(page_id: PageId) -> StorageError {
-        StorageError::Corruption(CorruptionError {
-            component: CorruptionComponent::InteriorPage,
-            page_id: Some(page_id),
-            kind: CorruptionKind::CellLengthOutOfBounds,
-        })
-    }
-
-    /// Returns the previous and next sibling pointers for a leaf page.
-    fn read_leaf_page_links(
-        &self,
-        page_id: PageId,
-    ) -> StorageResult<(Option<PageId>, Option<PageId>)> {
-        let pin = self.page_cache.fetch_page(page_id)?;
-        let page = pin.read()?;
-        let leaf = page.open::<Leaf>()?;
-        Ok((leaf.prev_page_id(), leaf.next_page_id()))
-    }
-
-    /// Returns the previous and next sibling pointers for an interior page.
-    fn read_interior_page_links(
-        &self,
-        page_id: PageId,
-    ) -> StorageResult<(Option<PageId>, Option<PageId>)> {
-        let pin = self.page_cache.fetch_page(page_id)?;
-        let page = pin.read()?;
-        let interior = page.open::<Interior>()?;
-        Ok((interior.prev_page_id(), interior.next_page_id()))
-    }
-
-    /// Materializes all leaf cells in slot order for rebalance planning.
-    fn read_leaf_cells(&self, page_id: PageId) -> StorageResult<Vec<LeafSplitCell>> {
-        let slot_count = self.raw_leaf_slot_count(page_id)?;
-        let mut cells = Vec::with_capacity(slot_count as usize);
-        for slot_index in 0..slot_count {
-            let (key, value) = read_leaf_cell(&self.page_cache, page_id, slot_index)?;
-            cells.push(LeafSplitCell { key, value });
-        }
-        Ok(cells)
-    }
-
-    /// Returns the largest key in a leaf page, or `None` when the page is empty.
-    fn read_leaf_max_key(&self, page_id: PageId) -> StorageResult<Option<Vec<u8>>> {
-        let slot_count = self.raw_leaf_slot_count(page_id)?;
-        if slot_count == 0 {
-            return Ok(None);
-        }
-
-        read_leaf_cell(&self.page_cache, page_id, slot_count - 1).map(|(key, _)| Some(key))
-    }
-
-    /// Reads child page ids from an interior page in logical left-to-right order.
-    fn read_interior_child_page_ids(&self, page_id: PageId) -> StorageResult<Vec<PageId>> {
-        let pin = self.page_cache.fetch_page(page_id)?;
-        let page = pin.read()?;
-        let interior = page.open::<Interior>()?;
-        let mut children = Vec::with_capacity(interior.slot_count() as usize + 1);
-        for slot_index in 0..interior.slot_count() {
-            let (left_child, _, _, _) = interior.cell_payload_parts(slot_index)?;
-            children.push(left_child);
-        }
-        children.push(interior.rightmost_child());
-        Ok(children)
-    }
-
-    /// Returns whether `child_page_id` is still linked from `parent_page_id`.
-    #[cfg_attr(not(test), allow(dead_code))]
-    fn interior_page_has_child(
-        &self,
-        parent_page_id: PageId,
-        child_page_id: PageId,
-    ) -> StorageResult<bool> {
-        let pin = self.page_cache.fetch_page(parent_page_id)?;
-        let page = pin.read()?;
-        let interior = page.open::<Interior>()?;
-        for slot_index in 0..interior.slot_count() {
-            let (left_child, _, _, _) = interior.cell_payload_parts(slot_index)?;
-            if left_child == child_page_id {
-                return Ok(true);
-            }
-        }
-        Ok(interior.rightmost_child() == child_page_id)
-    }
-
-    /// Returns the largest key reachable from the subtree rooted at `page_id`.
-    fn subtree_max_key(&self, page_id: PageId) -> StorageResult<Option<Vec<u8>>> {
-        let pin = self.page_cache.fetch_page(page_id)?;
-        let next = {
-            let page = pin.read()?;
-            match read_page_kind(page.page(), page_id)? {
-                PageKind::RawLeaf => {
-                    drop(page);
-                    drop(pin);
-                    return self.read_leaf_max_key(page_id);
-                }
-                PageKind::RawInterior => {
-                    let interior = page.open::<Interior>()?;
-                    interior.rightmost_child()
-                }
-            }
-        };
-        drop(pin);
-        self.subtree_max_key(next)
-    }
-
-    /// Collects ordered child entries for an interior page with refreshed max keys.
-    fn read_interior_child_entries(&self, page_id: PageId) -> StorageResult<Vec<ChildEntry>> {
-        let child_page_ids = self.read_interior_child_page_ids(page_id)?;
-        let mut children = Vec::with_capacity(child_page_ids.len());
-        for child_page_id in child_page_ids {
-            children.push(ChildEntry {
-                page_id: child_page_id,
-                max_key: self.subtree_max_key(child_page_id)?,
-            });
-        }
-        Ok(children)
-    }
-
-    /// Locates a child page within its parent's ordered child list.
-    fn child_index_in_parent(
-        &self,
-        parent_page_id: PageId,
-        child_page_id: PageId,
-    ) -> StorageResult<usize> {
-        let pin = self.page_cache.fetch_page(parent_page_id)?;
-        let page = pin.read()?;
-        let interior = page.open::<Interior>()?;
-        for slot_index in 0..interior.slot_count() {
-            let (left_child, _, _, _) = interior.cell_payload_parts(slot_index)?;
-            if left_child == child_page_id {
-                return Ok(slot_index as usize);
-            }
-        }
-        if interior.rightmost_child() == child_page_id {
-            return Ok(interior.slot_count() as usize);
-        }
-        Err(StorageError::Corruption(CorruptionError {
-            component: CorruptionComponent::InteriorPage,
-            page_id: Some(parent_page_id),
-            kind: CorruptionKind::CellLengthOutOfBounds,
-        }))
-    }
-
-    /// Reinitializes a leaf page with `cells` and updated sibling links.
-    fn rewrite_leaf_page(
-        &self,
-        page_id: PageId,
-        cells: &[LeafSplitCell],
-        prev_page_id: Option<PageId>,
-        next_page_id: Option<PageId>,
-    ) -> StorageResult<()> {
-        let pin = self.page_cache.fetch_page(page_id)?;
-        let mut guard = pin.write()?;
-        let mut leaf = RawLeaf::<Write<'_>>::initialize(guard.page_mut());
-        leaf.set_prev_page_id(prev_page_id);
-        leaf.set_next_page_id(next_page_id);
-        for cell in cells {
-            let slot_index = leaf.slot_count();
-            self.insert_leaf_payload_at(&mut leaf, slot_index, cell.key(), cell.value())?;
-        }
-        Ok(())
-    }
-
-    /// Reinitializes an interior page from ordered `children` and sibling links.
-    fn rewrite_interior_page(
-        &self,
-        page_id: PageId,
-        children: &[ChildEntry],
-        prev_page_id: Option<PageId>,
-        next_page_id: Option<PageId>,
-    ) -> StorageResult<()> {
-        let rightmost_child = children.last().ok_or({
-            StorageError::Corruption(CorruptionError {
-                component: CorruptionComponent::InteriorPage,
-                page_id: Some(page_id),
-                kind: CorruptionKind::CellLengthOutOfBounds,
-            })
-        })?;
-        let mut used_bytes = PageKind::RawInterior.header_size()
-            + (children.len() - 1) * page::format::SLOT_ENTRY_SIZE;
-        for child in &children[..children.len() - 1] {
-            let key = child
-                .max_key
-                .as_deref()
-                .ok_or_else(|| Self::missing_child_max_key_error(page_id))?;
-            used_bytes += self.interior_cell_local_size(key)?;
-        }
-        if used_bytes > page::format::USABLE_SPACE_END {
-            return Err(PageError::PageFull {
-                needed: used_bytes,
-                available: page::format::USABLE_SPACE_END,
-            }
-            .into());
-        }
-
-        let mut page_image = [0; PAGE_SIZE];
-        {
-            let mut interior = RawInterior::<Write<'_>>::initialize_with_rightmost(
-                &mut page_image,
-                rightmost_child.page_id,
-            );
-            interior.set_prev_page_id(prev_page_id);
-            interior.set_next_page_id(next_page_id);
-            for child in &children[..children.len() - 1] {
-                let key = child
-                    .max_key
-                    .as_deref()
-                    .ok_or_else(|| Self::missing_child_max_key_error(page_id))?;
-                let slot_index = interior.slot_count();
-                self.insert_interior_payload_at(&mut interior, slot_index, child.page_id, key)?;
-            }
-        }
-
-        let pin = self.page_cache.fetch_page(page_id)?;
-        let mut guard = pin.write()?;
-        *guard.page_mut() = page_image;
-        Ok(())
-    }
-
-    /// Returns whether an interior page already matches refreshed child maxima.
-    fn interior_page_matches_children(
-        &self,
-        page_id: PageId,
-        children: &[ChildEntry],
-    ) -> StorageResult<bool> {
-        let current_children = self.read_interior_child_page_ids(page_id)?;
-        if current_children.len() != children.len()
-            || current_children
-                .iter()
-                .zip(children)
-                .any(|(&current, desired)| current != desired.page_id)
-        {
-            return Ok(false);
-        }
-
-        for (slot_index, child) in children[..children.len() - 1].iter().enumerate() {
-            let expected_key = child
-                .max_key
-                .as_deref()
-                .ok_or_else(|| Self::missing_child_max_key_error(page_id))?;
-            let (_, actual_key) = read_interior_cell(&self.page_cache, page_id, slot_index as u16)?;
-            if actual_key != expected_key {
-                return Ok(false);
-            }
-        }
-
-        Ok(true)
-    }
-
-    /// Refreshes one interior page only when one of its separators changed.
-    fn refresh_interior_page_separators(&self, page_id: PageId) -> StorageResult<()> {
-        let children = self.read_interior_child_entries(page_id)?;
-        if self.interior_page_matches_children(page_id, &children)? {
-            return Ok(());
-        }
-
-        let (prev_page_id, next_page_id) = self.read_interior_page_links(page_id)?;
-        self.rewrite_interior_page(page_id, &children, prev_page_id, next_page_id)
-    }
-
-    /// Returns whether a leaf page is below the minimum occupancy target.
-    fn leaf_page_underoccupied(&self, page_id: PageId) -> StorageResult<bool> {
-        let pin = self.page_cache.fetch_page(page_id)?;
-        let page = pin.read()?;
-        let leaf = page.open::<Leaf>()?;
-        Ok(leaf.is_underoccupied()?)
-    }
-
-    /// Returns whether an interior page is below the minimum occupancy target.
-    fn interior_page_underoccupied(&self, page_id: PageId) -> StorageResult<bool> {
-        let pin = self.page_cache.fetch_page(page_id)?;
-        let page = pin.read()?;
-        let interior = page.open::<Interior>()?;
-        Ok(interior.is_underoccupied()?)
-    }
-
-    /// Returns whether a leaf rebuilt from `cells` would be underoccupied.
-    fn leaf_cells_underoccupied(cells: &[LeafSplitCell]) -> bool {
-        let occupied_variable_bytes = cells.len() * page::format::SLOT_ENTRY_SIZE
-            + cells.iter().map(LeafSplitCell::encoded_size).sum::<usize>();
-        let usable_variable_bytes =
-            page::format::USABLE_SPACE_END - PageKind::RawLeaf.header_size();
-        occupied_variable_bytes * 2 < usable_variable_bytes
-    }
-
-    /// Chooses a split index that keeps both leaf siblings fit and occupied.
-    fn choose_leaf_rebalance_split(cells: &[LeafSplitCell]) -> Option<usize> {
-        let total_cell_len = cells.iter().map(LeafSplitCell::encoded_size).sum::<usize>();
-        let mut left_cell_len = 0;
-        let mut best = None;
-
-        for split_index in 1..cells.len() {
-            left_cell_len += cells[split_index - 1].encoded_size();
-            if !Self::leaf_cells_fit(&cells[..split_index])
-                || !Self::leaf_cells_fit(&cells[split_index..])
-                || Self::leaf_cells_underoccupied(&cells[..split_index])
-                || Self::leaf_cells_underoccupied(&cells[split_index..])
-            {
-                continue;
-            }
-
-            let right_cell_len = total_cell_len - left_cell_len;
-            let imbalance = left_cell_len.abs_diff(right_cell_len);
-            if best.is_none_or(|(best_imbalance, _)| imbalance < best_imbalance) {
-                best = Some((imbalance, split_index));
-            }
-        }
-
-        best.map(|(_, split_index)| split_index)
-    }
-
-    /// Chooses a split index that keeps both leaf siblings within page capacity.
-    fn choose_leaf_fitting_split(cells: &[LeafSplitCell]) -> Option<usize> {
-        let total_cell_len = cells.iter().map(LeafSplitCell::encoded_size).sum::<usize>();
-        let mut left_cell_len = 0;
-        let mut best = None;
-
-        for split_index in 1..cells.len() {
-            left_cell_len += cells[split_index - 1].encoded_size();
-            if !Self::leaf_cells_fit(&cells[..split_index])
-                || !Self::leaf_cells_fit(&cells[split_index..])
-            {
-                continue;
-            }
-
-            let right_cell_len = total_cell_len - left_cell_len;
-            let imbalance = left_cell_len.abs_diff(right_cell_len);
-            if best.is_none_or(|(best_imbalance, _)| imbalance < best_imbalance) {
-                best = Some((imbalance, split_index));
-            }
-        }
-
-        best.map(|(_, split_index)| split_index)
-    }
-
-    /// Returns whether `children` can be encoded in one interior page.
-    fn interior_children_fit(children: &[ChildEntry]) -> bool {
-        if children.is_empty() {
-            return false;
-        }
-        let mut cell_bytes = 0;
-        for child in &children[..children.len() - 1] {
-            let Some(key) = child.max_key.as_ref() else {
-                return false;
-            };
-            cell_bytes += INTERIOR_CELL_PREFIX_SIZE + local_payload_len(key.len());
-        }
-        let used_bytes = PageKind::RawInterior.header_size()
-            + (children.len() - 1) * page::format::SLOT_ENTRY_SIZE
-            + cell_bytes;
-        used_bytes <= page::format::USABLE_SPACE_END
-    }
-
-    /// Returns whether an interior page rebuilt from `children` would be underoccupied.
-    fn interior_children_underoccupied(children: &[ChildEntry]) -> bool {
-        let mut cell_bytes = 0;
-        for child in &children[..children.len().saturating_sub(1)] {
-            let Some(key) = child.max_key.as_ref() else {
-                return true;
-            };
-            cell_bytes += INTERIOR_CELL_PREFIX_SIZE + local_payload_len(key.len());
-        }
-        let occupied_variable_bytes =
-            children.len().saturating_sub(1) * page::format::SLOT_ENTRY_SIZE + cell_bytes;
-        let usable_variable_bytes =
-            page::format::USABLE_SPACE_END - PageKind::RawInterior.header_size();
-        occupied_variable_bytes * 2 < usable_variable_bytes
-    }
-
-    /// Chooses a split index that keeps both interior siblings fit and occupied.
-    fn choose_interior_rebalance_split(children: &[ChildEntry]) -> Option<usize> {
-        let mut best = None;
-        for split_index in 1..children.len() {
-            let left = &children[..split_index];
-            let right = &children[split_index..];
-            if !Self::interior_children_fit(left)
-                || !Self::interior_children_fit(right)
-                || Self::interior_children_underoccupied(left)
-                || Self::interior_children_underoccupied(right)
-            {
-                continue;
-            }
-
-            let imbalance = split_index.abs_diff(children.len() - split_index);
-            if best.is_none_or(|(best_imbalance, _)| imbalance < best_imbalance) {
-                best = Some((imbalance, split_index));
-            }
-        }
-        best.map(|(_, split_index)| split_index)
-    }
-
-    /// Chooses a split index that keeps both interior siblings within page capacity.
-    fn choose_interior_fitting_split(children: &[ChildEntry]) -> Option<usize> {
-        let mut best = None;
-        for split_index in 1..children.len() {
-            let left = &children[..split_index];
-            let right = &children[split_index..];
-            if !Self::interior_children_fit(left) || !Self::interior_children_fit(right) {
-                continue;
-            }
-
-            let imbalance = split_index.abs_diff(children.len() - split_index);
-            if best.is_none_or(|(best_imbalance, _)| imbalance < best_imbalance) {
-                best = Some((imbalance, split_index));
-            }
-        }
-        best.map(|(_, split_index)| split_index)
-    }
-
-    /// Updates the previous-sibling link for a leaf page.
-    fn set_leaf_prev_page_id(
-        &self,
-        page_id: PageId,
-        prev_page_id: Option<PageId>,
-    ) -> StorageResult<()> {
-        let pin = self.page_cache.fetch_page(page_id)?;
-        let mut guard = pin.write()?;
-        let mut leaf = guard.open_mut::<Leaf>()?;
-        leaf.set_prev_page_id(prev_page_id);
-        Ok(())
-    }
-
-    /// Updates the previous-sibling link for an interior page.
-    fn set_interior_prev_page_id(
-        &self,
-        page_id: PageId,
-        prev_page_id: Option<PageId>,
-    ) -> StorageResult<()> {
-        let pin = self.page_cache.fetch_page(page_id)?;
-        let mut guard = pin.write()?;
-        let mut interior = guard.open_mut::<Interior>()?;
-        interior.set_prev_page_id(prev_page_id);
-        Ok(())
-    }
-
-    /// Removes `child_page_id` from `parent_page_id` and rewrites the parent.
-    fn remove_child_from_parent(
-        &self,
-        parent_page_id: PageId,
-        child_page_id: PageId,
-    ) -> StorageResult<()> {
-        let mut children = self.read_interior_child_entries(parent_page_id)?;
-        let child_index =
-            children.iter().position(|child| child.page_id == child_page_id).ok_or({
-                StorageError::Corruption(CorruptionError {
-                    component: CorruptionComponent::InteriorPage,
-                    page_id: Some(parent_page_id),
-                    kind: CorruptionKind::CellLengthOutOfBounds,
-                })
-            })?;
-        children.remove(child_index);
-        let (prev_page_id, next_page_id) = self.read_interior_page_links(parent_page_id)?;
-        self.rewrite_interior_page(parent_page_id, &children, prev_page_id, next_page_id)
-    }
-
-    /// Rewrites adjacent leaf siblings after redistributing their combined cells.
-    fn redistribute_leaf_pair(
-        &self,
-        left_page_id: PageId,
-        right_page_id: PageId,
-        cells: &[LeafSplitCell],
-        split_index: usize,
-    ) -> StorageResult<()> {
-        let (left_prev_page_id, _) = self.read_leaf_page_links(left_page_id)?;
-        let (_, right_next_page_id) = self.read_leaf_page_links(right_page_id)?;
-        self.rewrite_leaf_page(
-            left_page_id,
-            &cells[..split_index],
-            left_prev_page_id,
-            Some(right_page_id),
-        )?;
-        self.rewrite_leaf_page(
-            right_page_id,
-            &cells[split_index..],
-            Some(left_page_id),
-            right_next_page_id,
-        )
-    }
-
-    /// Merges two adjacent leaf pages into `survivor_page_id`.
-    fn merge_leaf_pages(
-        &self,
-        survivor_page_id: PageId,
-        removed_page_id: PageId,
-        cells: &[LeafSplitCell],
-    ) -> StorageResult<()> {
-        let (survivor_prev_page_id, _) = self.read_leaf_page_links(survivor_page_id)?;
-        let (_, removed_next_page_id) = self.read_leaf_page_links(removed_page_id)?;
-        self.rewrite_leaf_page(
-            survivor_page_id,
-            cells,
-            survivor_prev_page_id,
-            removed_next_page_id,
-        )?;
-        if let Some(next_page_id) = removed_next_page_id {
-            self.set_leaf_prev_page_id(next_page_id, Some(survivor_page_id))?;
-        }
-        Ok(())
-    }
-
-    fn sort_leaf_cells(cells: &mut [LeafSplitCell]) {
-        cells.sort_by(|left, right| left.key().cmp(right.key()));
-    }
-
-    /// Rebalances an underoccupied leaf against siblings.
-    ///
-    /// Returns `true` when a merge removed one child from the parent page.
-    fn rebalance_leaf_page(
-        &mut self,
-        leaf_page_id: PageId,
-        parent_page_id: PageId,
-    ) -> StorageResult<bool> {
-        let child_index = self.child_index_in_parent(parent_page_id, leaf_page_id)?;
-        let parent_children = self.read_interior_child_page_ids(parent_page_id)?;
-
-        if child_index > 0 {
-            let left_page_id = parent_children[child_index - 1];
-            let mut cells = self.read_leaf_cells(left_page_id)?;
-            cells.extend(self.read_leaf_cells(leaf_page_id)?);
-            Self::sort_leaf_cells(&mut cells);
-            if let Some(split_index) = Self::choose_leaf_rebalance_split(&cells) {
-                self.redistribute_leaf_pair(left_page_id, leaf_page_id, &cells, split_index)?;
-                return Ok(false);
-            }
-        }
-
-        if child_index + 1 < parent_children.len() {
-            let right_page_id = parent_children[child_index + 1];
-            let mut cells = self.read_leaf_cells(leaf_page_id)?;
-            cells.extend(self.read_leaf_cells(right_page_id)?);
-            Self::sort_leaf_cells(&mut cells);
-            if let Some(split_index) = Self::choose_leaf_rebalance_split(&cells) {
-                self.redistribute_leaf_pair(leaf_page_id, right_page_id, &cells, split_index)?;
-                return Ok(false);
-            }
-        }
-
-        if child_index > 0 {
-            let left_page_id = parent_children[child_index - 1];
-            let mut cells = self.read_leaf_cells(left_page_id)?;
-            cells.extend(self.read_leaf_cells(leaf_page_id)?);
-            Self::sort_leaf_cells(&mut cells);
-            if Self::leaf_cells_fit(&cells) {
-                self.merge_leaf_pages(left_page_id, leaf_page_id, &cells)?;
-                self.remove_child_from_parent(parent_page_id, leaf_page_id)?;
-                self.set_page_state(left_page_id);
-                return Ok(true);
-            }
-            if let Some(split_index) = Self::choose_leaf_fitting_split(&cells) {
-                self.redistribute_leaf_pair(left_page_id, leaf_page_id, &cells, split_index)?;
-                return Ok(false);
-            }
-        }
-
-        if child_index + 1 < parent_children.len() {
-            let right_page_id = parent_children[child_index + 1];
-            let mut cells = self.read_leaf_cells(leaf_page_id)?;
-            cells.extend(self.read_leaf_cells(right_page_id)?);
-            Self::sort_leaf_cells(&mut cells);
-            if Self::leaf_cells_fit(&cells) {
-                self.merge_leaf_pages(leaf_page_id, right_page_id, &cells)?;
-                self.remove_child_from_parent(parent_page_id, right_page_id)?;
-                self.set_page_state(leaf_page_id);
-                return Ok(true);
-            }
-            if let Some(split_index) = Self::choose_leaf_fitting_split(&cells) {
-                self.redistribute_leaf_pair(leaf_page_id, right_page_id, &cells, split_index)?;
-                return Ok(false);
-            }
-        }
-
-        Ok(false)
-    }
-
-    /// Rewrites adjacent interior siblings after redistributing their children.
-    fn redistribute_interior_pair(
-        &self,
-        left_page_id: PageId,
-        right_page_id: PageId,
-        children: &[ChildEntry],
-        split_index: usize,
-    ) -> StorageResult<()> {
-        let (left_prev_page_id, _) = self.read_interior_page_links(left_page_id)?;
-        let (_, right_next_page_id) = self.read_interior_page_links(right_page_id)?;
-        self.rewrite_interior_page(
-            left_page_id,
-            &children[..split_index],
-            left_prev_page_id,
-            Some(right_page_id),
-        )?;
-        self.rewrite_interior_page(
-            right_page_id,
-            &children[split_index..],
-            Some(left_page_id),
-            right_next_page_id,
-        )
-    }
-
-    /// Merges two adjacent interior pages into `survivor_page_id`.
-    fn merge_interior_pages(
-        &self,
-        survivor_page_id: PageId,
-        removed_page_id: PageId,
-        children: &[ChildEntry],
-    ) -> StorageResult<()> {
-        let (survivor_prev_page_id, _) = self.read_interior_page_links(survivor_page_id)?;
-        let (_, removed_next_page_id) = self.read_interior_page_links(removed_page_id)?;
-        self.rewrite_interior_page(
-            survivor_page_id,
-            children,
-            survivor_prev_page_id,
-            removed_next_page_id,
-        )?;
-        if let Some(next_page_id) = removed_next_page_id {
-            self.set_interior_prev_page_id(next_page_id, Some(survivor_page_id))?;
-        }
-        Ok(())
-    }
-
-    /// Splits an existing interior page after refreshed separators no longer fit.
-    fn split_existing_interior_page(
-        &mut self,
-        page_id: PageId,
-        children: &[ChildEntry],
-    ) -> StorageResult<PendingSplit> {
-        let (prev_page_id, next_page_id) = self.read_interior_page_links(page_id)?;
-        let (right_page_id, right_page_guard) = self.page_cache.new_page()?;
-        drop(right_page_guard);
-
-        let split_index = Self::choose_interior_fitting_split(children)
-            .ok_or(PageError::PageFull { needed: PAGE_SIZE + 1, available: PAGE_SIZE })?;
-        let (left_children, right_children) = children.split_at(split_index);
-
-        let propagated_separator = left_children
-            .last()
-            .and_then(|child| child.max_key.clone())
-            .ok_or_else(|| Self::missing_child_max_key_error(page_id))?;
-
-        self.rewrite_interior_page(page_id, left_children, prev_page_id, Some(right_page_id))?;
-        self.rewrite_interior_page(right_page_id, right_children, Some(page_id), next_page_id)?;
-
-        if let Some(next_page_id) = next_page_id {
-            self.set_interior_prev_page_id(next_page_id, Some(right_page_id))?;
-        }
-
-        Ok(PendingSplit { separator: propagated_separator, left_page_id: page_id, right_page_id })
-    }
-
-    /// Rebalances an underoccupied interior page against siblings.
-    ///
-    /// Returns `true` when a merge removed one child from the parent page.
-    fn rebalance_interior_page(
-        &self,
-        interior_page_id: PageId,
-        parent_page_id: PageId,
-    ) -> StorageResult<bool> {
-        let child_index = self.child_index_in_parent(parent_page_id, interior_page_id)?;
-        let parent_children = self.read_interior_child_page_ids(parent_page_id)?;
-
-        if child_index > 0 {
-            let left_page_id = parent_children[child_index - 1];
-            let mut children = self.read_interior_child_entries(left_page_id)?;
-            children.extend(self.read_interior_child_entries(interior_page_id)?);
-            if let Some(split_index) = Self::choose_interior_rebalance_split(&children) {
-                self.redistribute_interior_pair(
-                    left_page_id,
-                    interior_page_id,
-                    &children,
-                    split_index,
-                )?;
-                return Ok(false);
-            }
-        }
-
-        if child_index + 1 < parent_children.len() {
-            let right_page_id = parent_children[child_index + 1];
-            let mut children = self.read_interior_child_entries(interior_page_id)?;
-            children.extend(self.read_interior_child_entries(right_page_id)?);
-            if let Some(split_index) = Self::choose_interior_rebalance_split(&children) {
-                self.redistribute_interior_pair(
-                    interior_page_id,
-                    right_page_id,
-                    &children,
-                    split_index,
-                )?;
-                return Ok(false);
-            }
-        }
-
-        if child_index > 0 {
-            let left_page_id = parent_children[child_index - 1];
-            let mut children = self.read_interior_child_entries(left_page_id)?;
-            children.extend(self.read_interior_child_entries(interior_page_id)?);
-            if Self::interior_children_fit(&children) {
-                self.merge_interior_pages(left_page_id, interior_page_id, &children)?;
-                self.remove_child_from_parent(parent_page_id, interior_page_id)?;
-                return Ok(true);
-            }
-            if let Some(split_index) = Self::choose_interior_fitting_split(&children) {
-                self.redistribute_interior_pair(
-                    left_page_id,
-                    interior_page_id,
-                    &children,
-                    split_index,
-                )?;
-                return Ok(false);
-            }
-        }
-
-        if child_index + 1 < parent_children.len() {
-            let right_page_id = parent_children[child_index + 1];
-            let mut children = self.read_interior_child_entries(interior_page_id)?;
-            children.extend(self.read_interior_child_entries(right_page_id)?);
-            if Self::interior_children_fit(&children) {
-                self.merge_interior_pages(interior_page_id, right_page_id, &children)?;
-                self.remove_child_from_parent(parent_page_id, right_page_id)?;
-                return Ok(true);
-            }
-            if let Some(split_index) = Self::choose_interior_fitting_split(&children) {
-                self.redistribute_interior_pair(
-                    interior_page_id,
-                    right_page_id,
-                    &children,
-                    split_index,
-                )?;
-                return Ok(false);
-            }
-        }
-
-        Ok(false)
-    }
-
-    /// Replaces an empty interior root with its sole child while preserving the root page id.
-    fn shrink_root_if_empty(&mut self) -> StorageResult<()> {
-        let root_page_id = self.root_page_id();
-        let pin = self.page_cache.fetch_page(root_page_id)?;
-        let child_page_id = {
-            let page = pin.read()?;
-            match read_page_kind(page.page(), root_page_id)? {
-                PageKind::RawLeaf => return Ok(()),
-                PageKind::RawInterior => {
-                    let interior = page.open::<Interior>()?;
-                    if interior.slot_count() > 0 {
-                        return Ok(());
-                    }
-                    interior.rightmost_child()
-                }
-            }
-        };
-
-        let child_pin = self.page_cache.fetch_page(child_page_id)?;
-        let child_snapshot = {
-            let child_page = child_pin.read()?;
-            *child_page.page()
-        };
-
-        let mut root_guard = pin.write()?;
-        *root_guard.page_mut() = child_snapshot;
-        drop(root_guard);
-
-        self.clear_root_sibling_links(root_page_id)?;
-        self.set_page_state(root_page_id);
-        Ok(())
-    }
-
-    fn clear_root_sibling_links(&self, root_page_id: PageId) -> StorageResult<()> {
-        let pin = self.page_cache.fetch_page(root_page_id)?;
-        let mut guard = pin.write()?;
-        match read_page_kind(guard.page(), root_page_id)? {
-            PageKind::RawLeaf => {
-                let mut leaf = guard.open_mut::<Leaf>()?;
-                leaf.set_prev_page_id(None);
-                leaf.set_next_page_id(None);
-            }
-            PageKind::RawInterior => {
-                let mut interior = guard.open_mut::<Interior>()?;
-                interior.set_prev_page_id(None);
-                interior.set_next_page_id(None);
-            }
-        }
-        Ok(())
-    }
-
-    /// Refreshes separators along the still-reachable delete path.
-    #[cfg_attr(not(test), allow(dead_code))]
-    fn refresh_path_separators(&self, tree_path: &[PathFrame]) -> StorageResult<()> {
-        if tree_path.is_empty() {
-            return Ok(());
-        }
-
-        let mut reachable_depth = 0;
-        for (depth, frame) in tree_path.iter().enumerate() {
-            let is_reachable = if depth == 0 {
-                frame.page_id == self.root_page_id()
-            } else {
-                self.interior_page_has_child(tree_path[depth - 1].page_id, frame.page_id)?
-            };
-            if !is_reachable {
-                break;
-            }
-            reachable_depth += 1;
-        }
-
-        for frame in tree_path[..reachable_depth].iter().rev() {
-            self.refresh_interior_page_separators(frame.page_id)?;
-        }
-
-        Ok(())
-    }
-
-    /// Refreshes separators throughout the reachable subtree rooted at the current root.
-    fn refresh_subtree_separators(&mut self) -> StorageResult<()> {
-        loop {
-            let root_page_id = self.root_page_id();
-            let Some(pending) = self.refresh_subtree_separators_once(root_page_id)? else {
-                return Ok(());
-            };
-            self.install_new_root(pending)?;
-        }
-    }
-
-    /// Refreshes one subtree and returns a split that must be inserted by its parent.
-    fn refresh_subtree_separators_once(
-        &mut self,
-        page_id: PageId,
-    ) -> StorageResult<Option<PendingSplit>> {
-        let pin = self.page_cache.fetch_page(page_id)?;
-        let is_interior = {
-            let page = pin.read()?;
-            match read_page_kind(page.page(), page_id)? {
-                PageKind::RawLeaf => false,
-                PageKind::RawInterior => true,
-            }
-        };
-        drop(pin);
-        if !is_interior {
-            return Ok(None);
-        }
-
-        let mut child_index = 0;
-        loop {
-            let child_page_ids = self.read_interior_child_page_ids(page_id)?;
-            let Some(&child_page_id) = child_page_ids.get(child_index) else {
-                break;
-            };
-
-            if let Some(pending) = self.refresh_subtree_separators_once(child_page_id)? {
-                let child_ref = if child_index + 1 == child_page_ids.len() {
-                    ChildSlotRef::Rightmost
-                } else {
-                    ChildSlotRef::Slot(child_index as u16)
-                };
-                let parent_frame = PathFrame { page_id, child_ref };
-                if let Some(parent_pending) = self.insert_into_parent(parent_frame, pending)? {
-                    return Ok(Some(parent_pending));
-                }
-                child_index += 2;
-            } else {
-                child_index += 1;
-            }
-        }
-
-        let children = self.read_interior_child_entries(page_id)?;
-        if self.interior_page_matches_children(page_id, &children)? {
-            return Ok(None);
-        }
-
-        if Self::interior_children_fit(&children) {
-            let (prev_page_id, next_page_id) = self.read_interior_page_links(page_id)?;
-            self.rewrite_interior_page(page_id, &children, prev_page_id, next_page_id)?;
-            return Ok(None);
-        }
-
-        self.split_existing_interior_page(page_id, &children).map(Some)
-    }
-
-    /// Runs post-delete rebalancing from the modified leaf toward the root.
-    fn rebalance_after_leaf_delete(
-        &mut self,
-        leaf_page_id: PageId,
-        tree_path: &[PathFrame],
-    ) -> StorageResult<()> {
-        if tree_path.is_empty() {
-            return Ok(());
-        }
-        if !self.leaf_page_underoccupied(leaf_page_id)? {
-            return Ok(());
-        }
-
-        let mut depth = tree_path.len() - 1;
-        let parent_page_id = tree_path[depth].page_id;
-        if !self.rebalance_leaf_page(leaf_page_id, parent_page_id)? {
-            return Ok(());
-        }
-
-        loop {
-            let page_id = tree_path[depth].page_id;
-            if page_id == self.root_page_id() {
-                self.shrink_root_if_empty()?;
-                return Ok(());
-            }
-            if !self.interior_page_underoccupied(page_id)? {
-                return Ok(());
-            }
-            if depth == 0 {
-                return Ok(());
-            }
-
-            let parent_page_id = tree_path[depth - 1].page_id;
-            if !self.rebalance_interior_page(page_id, parent_page_id)? {
-                return Ok(());
-            }
-            depth -= 1;
-        }
-    }
-
-    /// Inserts a new raw key/value record into the tree.
-    pub fn insert(&mut self, key: &[u8], value: &[u8]) -> StorageResult<()> {
-        let (leaf_page_id, tree_path) = self.leaf_page_path_for_key(key)?;
-        let slot_index = match self.search_leaf_slot(leaf_page_id, key)? {
-            SearchResult::Found(_) => return Err(PageError::DuplicateKey.into()),
-            SearchResult::InsertAt(slot_index) => slot_index,
-        };
-        let leaf_pin_guard = self.page_cache.fetch_page(leaf_page_id)?;
-        let mut leaf_guard = leaf_pin_guard.write()?;
-        let has_capacity = {
-            let page = leaf_guard.open::<Leaf>()?;
-            let needed = self.leaf_cell_local_size(key, value)? + page::format::SLOT_ENTRY_SIZE;
-            page.total_reclaimable_space()? >= needed
-        };
-
-        if has_capacity {
-            {
-                let mut page = leaf_guard.open_mut::<Leaf>()?;
-                let slot_index = self.insert_leaf_payload_at(&mut page, slot_index, key, value)?;
-                self.set_positioned_state(leaf_page_id, slot_index);
-            }
-            drop(leaf_guard);
-            drop(leaf_pin_guard);
-            self.refresh_subtree_separators()?;
-            return Ok(());
-        }
-
-        let pending =
-            self.insert_with_leaf_page_split(leaf_page_id, &mut leaf_guard, key, value)?;
-        drop(leaf_guard);
-        drop(leaf_pin_guard);
-        self.propagate_split(&tree_path, pending)?;
-        self.refresh_subtree_separators()
-    }
-
-    /// Replaces the value stored for an existing `key`.
-    pub fn update(&mut self, key: &[u8], value: &[u8]) -> StorageResult<()> {
-        let (leaf_page_id, tree_path) = self.leaf_page_path_for_key(key)?;
-        let slot_index = match self.search_leaf_slot(leaf_page_id, key)? {
-            SearchResult::Found(slot_index) => slot_index,
-            SearchResult::InsertAt(_) => return Err(PageError::KeyNotFound.into()),
-        };
-        let leaf_pin_guard = self.page_cache.fetch_page(leaf_page_id)?;
-        let mut leaf_guard = leaf_pin_guard.write()?;
-        let has_capacity = {
-            let page = leaf_guard.open::<Leaf>()?;
-            let old_len = page.cell_len(slot_index)?;
-            let needed = self.leaf_cell_local_size(key, value)?;
-            page.total_reclaimable_space()? + old_len >= needed
-        };
-
-        if has_capacity {
-            {
-                let mut page = leaf_guard.open_mut::<Leaf>()?;
-                let slot_index = self.update_leaf_payload_at(&mut page, slot_index, key, value)?;
-                self.set_positioned_state(leaf_page_id, slot_index);
-            }
-            drop(leaf_guard);
-            drop(leaf_pin_guard);
-            self.refresh_path_separators(&tree_path)?;
-            return Ok(());
-        }
-
-        let pending =
-            self.update_with_leaf_page_split(leaf_page_id, &mut leaf_guard, slot_index, value)?;
-        drop(leaf_guard);
-        drop(leaf_pin_guard);
-        self.propagate_split(&tree_path, pending)?;
-        self.refresh_subtree_separators()
-    }
-
-    /// Deletes the record identified by `key`.
-    pub fn delete(&mut self, key: &[u8]) -> StorageResult<()> {
-        let (leaf_page_id, tree_path) = self.leaf_page_path_for_key(key)?;
-        {
-            let leaf_pin_guard = self.page_cache.fetch_page(leaf_page_id)?;
-            let mut leaf_guard = leaf_pin_guard.write()?;
-            let mut page = leaf_guard.open_mut::<Leaf>()?;
-            page.delete(key)?;
-        }
-
-        self.set_page_state(leaf_page_id);
-        self.rebalance_after_leaf_delete(leaf_page_id, &tree_path)?;
-        self.shrink_root_if_empty()?;
-        self.refresh_subtree_separators()?;
-        Ok(())
-    }
-
-    /// Positions the cursor on the first record whose key is greater than or
-    /// equal to `key`.
-    pub fn seek_to_key(&mut self, key: &[u8]) -> StorageResult<bool> {
-        let page_id = self.leaf_page_for_key(key)?;
-        let pin = self.page_cache.fetch_page(page_id)?;
-        let seek = {
-            let page = pin.read()?;
-            let leaf = page.open::<Leaf>()?;
-            let bound = self.lower_bound_leaf_slot(page_id, key)?;
-            match bound {
-                BoundResult::At(slot_index) => LeafSeek::Positioned(slot_index),
-                BoundResult::PastEnd => match leaf.next_page_id() {
-                    Some(next_page_id) => LeafSeek::Advance(next_page_id),
-                    None => LeafSeek::Exhausted,
-                },
-            }
-        };
-
-        match seek {
-            LeafSeek::Positioned(slot_index) => {
-                self.set_positioned_state(page_id, slot_index);
-                Ok(true)
-            }
-            LeafSeek::Advance(next_page_id) => self
-                .edge_record_from_leaf(next_page_id, ScanDirection::Forward)
-                .map(|record| record.is_some()),
-            LeafSeek::Exhausted => {
-                self.set_exhausted_state();
-                Ok(false)
-            }
-        }
-    }
-
-    /// Positions the cursor on the smallest key in the tree.
-    pub fn seek_to_first(&mut self) -> StorageResult<bool> {
-        let leaf_page_id = self.descend_to_first_leaf_from(self.root_page_id())?;
-        self.edge_record_from_leaf(leaf_page_id, ScanDirection::Forward)
-            .map(|record| record.is_some())
-    }
-
-    /// Reads the currently selected record, if any.
-    pub fn current(&self) -> StorageResult<Option<Record<S>>> {
-        match self.state {
-            CursorState::Positioned { page_id, slot_index } => {
-                self.record_at(page_id, slot_index).map(Some)
-            }
-            CursorState::Page { .. } | CursorState::Exhausted => Ok(None),
-        }
-    }
-
-    /// Reads the currently selected record as a stable owned snapshot, if any.
-    pub fn current_owned(&self) -> StorageResult<Option<OwnedRecord>> {
-        self.current()?.map(|record| record.to_owned_record()).transpose()
-    }
-
-    /// Advances to the next record in sorted key order.
-    pub fn next_record(&mut self) -> StorageResult<Option<Record<S>>> {
-        self.step_record(ScanDirection::Forward)
-    }
-
-    /// Advances to the next record and returns a stable owned snapshot.
-    pub fn next_owned_record(&mut self) -> StorageResult<Option<OwnedRecord>> {
-        self.next_record()?.map(|record| record.to_owned_record()).transpose()
-    }
-
-    /// Moves to the previous record in sorted key order.
-    pub fn prev_record(&mut self) -> StorageResult<Option<Record<S>>> {
-        self.step_record(ScanDirection::Backward)
-    }
-
-    /// Moves to the previous record and returns a stable owned snapshot.
-    pub fn prev_owned_record(&mut self) -> StorageResult<Option<OwnedRecord>> {
-        self.prev_record()?.map(|record| record.to_owned_record()).transpose()
-    }
-
-    /// Bubbles one pending split up the recorded tree path until it lands.
-    fn propagate_split(
-        &mut self,
-        tree_path: &[PathFrame],
-        mut pending: PendingSplit,
-    ) -> StorageResult<()> {
-        for &parent_frame in tree_path.iter().rev() {
-            match self.insert_into_parent(parent_frame, pending)? {
-                Some(next_pending) => pending = next_pending,
-                None => return Ok(()),
-            }
-        }
-
-        self.install_new_root(pending)
-    }
-
-    /// Re-points the parent-side child reference after inserting a separator.
-    fn update_interior_child_ref(
-        interior_page: &mut RawInterior<Write<'_>>,
-        child_ref: ChildSlotRef,
-        inserted_slot_index: u16,
-        child_page_id: PageId,
-    ) -> StorageResult<()> {
-        match child_ref {
-            ChildSlotRef::Slot(slot_index) => {
-                let updated_slot_index =
-                    if inserted_slot_index <= slot_index { slot_index + 1 } else { slot_index };
-                let mut cell = interior_page.cell_mut(updated_slot_index)?;
-                cell.set_left_child(child_page_id)?;
-            }
-            ChildSlotRef::Rightmost => {
-                interior_page.set_rightmost_child(child_page_id);
-            }
-        }
-        Ok(())
-    }
-
-    /// Inserts one promoted separator into an interior page or reports another split.
-    fn insert_into_parent(
-        &mut self,
-        parent_frame: PathFrame,
-        pending: PendingSplit,
-    ) -> StorageResult<Option<PendingSplit>> {
-        let insert_slot_index = match parent_frame.child_ref {
-            ChildSlotRef::Slot(slot_index) => slot_index,
-            ChildSlotRef::Rightmost => self.raw_interior_slot_count(parent_frame.page_id)?,
-        };
-        let interior_page_guard = self.page_cache.fetch_page(parent_frame.page_id)?;
-        let mut interior_guard = interior_page_guard.write()?;
-        let has_capacity = {
-            let page = interior_guard.open::<Interior>()?;
-            let needed =
-                self.interior_cell_local_size(&pending.separator)? + page::format::SLOT_ENTRY_SIZE;
-            page.total_reclaimable_space()? >= needed
-        };
-
-        if has_capacity {
-            let mut interior_page = interior_guard.open_mut::<Interior>()?;
-            let inserted_slot_index = self.insert_interior_payload_at(
-                &mut interior_page,
-                insert_slot_index,
-                pending.left_page_id,
-                &pending.separator,
-            )?;
-            Self::update_interior_child_ref(
-                &mut interior_page,
-                parent_frame.child_ref,
-                inserted_slot_index,
-                pending.right_page_id,
-            )
-            .map(|()| None)
-        } else {
-            drop(interior_guard);
-            drop(interior_page_guard);
-            self.insert_with_interior_page_split(parent_frame, pending).map(Some)
-        }
-    }
-
-    /// Returns whether the provided leaf cells fit into one leaf page.
-    fn leaf_cells_fit(cells: &[LeafSplitCell]) -> bool {
-        let used_bytes = PageKind::RawLeaf.header_size()
-            + cells.len() * page::format::SLOT_ENTRY_SIZE
-            + cells.iter().map(LeafSplitCell::encoded_size).sum::<usize>();
-        used_bytes <= page::format::USABLE_SPACE_END
-    }
-
-    /// Chooses the leaf split point with the smallest byte imbalance.
-    fn choose_leaf_split_index(cells: &[LeafSplitCell]) -> StorageResult<usize> {
-        debug_assert!(cells.len() >= 2, "leaf splits need at least two cells");
-
-        let total_cell_len = cells.iter().map(LeafSplitCell::encoded_size).sum::<usize>();
-        let mut left_cell_len = 0;
-        let mut best = None;
-
-        for split_index in 1..cells.len() {
-            left_cell_len += cells[split_index - 1].encoded_size();
-            if !Self::leaf_cells_fit(&cells[..split_index])
-                || !Self::leaf_cells_fit(&cells[split_index..])
-            {
-                continue;
-            }
-
-            let right_cell_len = total_cell_len - left_cell_len;
-            let imbalance = left_cell_len.abs_diff(right_cell_len);
-            let is_better = match best {
-                Some((best_imbalance, best_left_cell_len, _)) => {
-                    imbalance < best_imbalance
-                        || (imbalance == best_imbalance && left_cell_len > best_left_cell_len)
-                }
-                None => true,
-            };
-
-            if is_better {
-                best = Some((imbalance, left_cell_len, split_index));
-            }
-        }
-
-        match best {
-            Some((_, _, split_index)) => Ok(split_index),
-            None => Err(PageError::PageFull {
-                needed: total_cell_len,
-                available: total_cell_len.saturating_sub(1),
-            }
-            .into()),
-        }
-    }
-
-    /// Materializes all leaf cells from a stable page snapshot.
-    fn read_leaf_cells_from_snapshot(
-        &self,
-        leaf_page_id: PageId,
-        leaf_snapshot_bytes: &[u8; PAGE_SIZE],
-        leaf_snapshot: &RawLeaf<Read<'_>>,
-    ) -> StorageResult<Vec<LeafSplitCell>> {
-        let mut cells = Vec::with_capacity(leaf_snapshot.slot_count() as usize);
-        for slot_index in 0..leaf_snapshot.slot_count() {
-            let (key_len, value_len, first_overflow_page_id, inline_range) =
-                leaf_snapshot.cell_payload_parts(slot_index)?;
-            let mut payload = materialize_payload(
-                &self.page_cache,
-                leaf_page_id,
-                &leaf_snapshot_bytes[inline_range],
-                first_overflow_page_id,
-                key_len + value_len,
-            )?;
-            if payload.len() < key_len {
-                return Err(cell_corruption(leaf_page_id, CorruptionKind::CellLengthOutOfBounds));
-            }
-            let value = payload.split_off(key_len);
-            cells.push(LeafSplitCell { key: payload, value });
-        }
-        Ok(cells)
-    }
-
-    /// Rebuilds a split leaf pair from ordered materialized cells.
-    fn split_leaf_cells(
-        &mut self,
-        leaf_page_id: PageId,
-        leaf_guard: &mut PageWriteGuard<'_>,
-        prev_page_id: Option<PageId>,
-        next_page_id: Option<PageId>,
-        cells: &[LeafSplitCell],
-        target_key: &[u8],
-    ) -> StorageResult<PendingSplit> {
-        let split_index = Self::choose_leaf_split_index(cells)?;
-        let (left_cells, right_cells) = cells.split_at(split_index);
-
-        let (right_page_id, right_page_guard) = self.page_cache.new_page()?;
-        let mut right_guard = right_page_guard.write()?;
-        let mut right_page = RawLeaf::<Write<'_>>::initialize(right_guard.page_mut());
-
-        let mut leaf_page = RawLeaf::<Write<'_>>::initialize(leaf_guard.page_mut());
-        leaf_page.set_prev_page_id(prev_page_id);
-        leaf_page.set_next_page_id(Some(right_page_id));
-        right_page.set_prev_page_id(Some(leaf_page_id));
-        right_page.set_next_page_id(next_page_id);
-
-        for cell in left_cells {
-            let slot_index = leaf_page.slot_count();
-            self.insert_leaf_payload_at(&mut leaf_page, slot_index, cell.key(), cell.value())?;
-        }
-        for cell in right_cells {
-            let slot_index = right_page.slot_count();
-            self.insert_leaf_payload_at(&mut right_page, slot_index, cell.key(), cell.value())?;
-        }
-
-        if let Some(next_page_id) = next_page_id {
-            let next_page_guard = self.page_cache.fetch_page(next_page_id)?;
-            let mut next_guard = next_page_guard.write()?;
-            let mut next_page = next_guard.open_mut::<Leaf>()?;
-            next_page.set_prev_page_id(Some(right_page_id));
-        }
-
-        let separator =
-            left_cells.last().expect("leaf split must leave a non-empty left page").key().to_vec();
-
-        let target_page_id =
-            if target_key <= separator.as_slice() { leaf_page_id } else { right_page_id };
-        let target_cells = if target_page_id == leaf_page_id { left_cells } else { right_cells };
-        let target_slot_index = target_cells
-            .iter()
-            .position(|cell| cell.key() == target_key)
-            .expect("leaf split must retain the target key") as u16;
-        self.set_positioned_state(target_page_id, target_slot_index);
-
-        Ok(PendingSplit { separator, left_page_id: leaf_page_id, right_page_id })
-    }
-
-    /// Splits a full leaf page while inserting a new key/value cell.
-    fn insert_with_leaf_page_split(
-        &mut self,
-        leaf_page_id: PageId,
-        leaf_guard: &mut PageWriteGuard<'_>,
-        key: &[u8],
-        value: &[u8],
-    ) -> StorageResult<PendingSplit> {
-        let leaf_snapshot_bytes = *leaf_guard.page();
-        let leaf_snapshot = RawLeaf::<Read<'_>>::open(&leaf_snapshot_bytes)?;
-
-        let prev_page_id = leaf_snapshot.prev_page_id();
-        let next_page_id = leaf_snapshot.next_page_id();
-        let mut cells =
-            self.read_leaf_cells_from_snapshot(leaf_page_id, &leaf_snapshot_bytes, &leaf_snapshot)?;
-
-        let idx = match cells.binary_search_by(|cell| cell.key().cmp(key)) {
-            Ok(_) => return Err(PageError::DuplicateKey.into()),
-            Err(insert_index) => insert_index,
-        };
-        cells.insert(idx, LeafSplitCell { key: key.to_vec(), value: value.to_vec() });
-
-        self.split_leaf_cells(leaf_page_id, leaf_guard, prev_page_id, next_page_id, &cells, key)
-    }
-
-    /// Splits a full leaf page while replacing one existing cell value.
-    fn update_with_leaf_page_split(
-        &mut self,
-        leaf_page_id: PageId,
-        leaf_guard: &mut PageWriteGuard<'_>,
-        slot_index: u16,
-        value: &[u8],
-    ) -> StorageResult<PendingSplit> {
-        let leaf_snapshot_bytes = *leaf_guard.page();
-        let leaf_snapshot = RawLeaf::<Read<'_>>::open(&leaf_snapshot_bytes)?;
-
-        let prev_page_id = leaf_snapshot.prev_page_id();
-        let next_page_id = leaf_snapshot.next_page_id();
-        let mut cells =
-            self.read_leaf_cells_from_snapshot(leaf_page_id, &leaf_snapshot_bytes, &leaf_snapshot)?;
-        let target = cells
-            .get_mut(slot_index as usize)
-            .ok_or_else(|| cell_corruption(leaf_page_id, CorruptionKind::CellLengthOutOfBounds))?;
-        target.value = value.to_vec();
-        let target_key = target.key.clone();
-
-        self.split_leaf_cells(
-            leaf_page_id,
-            leaf_guard,
-            prev_page_id,
-            next_page_id,
-            &cells,
-            &target_key,
-        )
-    }
-
-    /// Splits a full interior page while inserting a propagated separator.
-    fn insert_with_interior_page_split(
-        &mut self,
-        parent_frame: PathFrame,
-        pending: PendingSplit,
-    ) -> StorageResult<PendingSplit> {
-        let PendingSplit { separator, left_page_id, right_page_id: incoming_right_page_id } =
-            pending;
-        let (prev_page_id, next_page_id) = self.read_interior_page_links(parent_frame.page_id)?;
-        let mut children = self.read_interior_child_entries(parent_frame.page_id)?;
-        let child_index = match parent_frame.child_ref {
-            ChildSlotRef::Slot(slot_index) => slot_index as usize,
-            ChildSlotRef::Rightmost => children.len() - 1,
-        };
-        let original_max_key = children[child_index].max_key.take();
-        children[child_index] = ChildEntry { page_id: left_page_id, max_key: Some(separator) };
-        children.insert(
-            child_index + 1,
-            ChildEntry { page_id: incoming_right_page_id, max_key: original_max_key },
-        );
-
-        let (right_page_id, right_page_guard) = self.page_cache.new_page()?;
-        drop(right_page_guard);
-
-        let split_index = Self::choose_interior_fitting_split(&children)
-            .ok_or(PageError::PageFull { needed: PAGE_SIZE + 1, available: PAGE_SIZE })?;
-        let (left_children, right_children) = children.split_at(split_index);
-
-        let propagated_separator = left_children
-            .last()
-            .and_then(|child| child.max_key.clone())
-            .ok_or_else(|| Self::missing_child_max_key_error(parent_frame.page_id))?;
-
-        self.rewrite_interior_page(
-            parent_frame.page_id,
-            left_children,
-            prev_page_id,
-            Some(right_page_id),
-        )?;
-        self.rewrite_interior_page(
-            right_page_id,
-            right_children,
-            Some(parent_frame.page_id),
-            next_page_id,
-        )?;
-
-        if let Some(next_page_id) = next_page_id {
-            let next_page_guard = self.page_cache.fetch_page(next_page_id)?;
-            let mut next_guard = next_page_guard.write()?;
-            let mut next_page = next_guard.open_mut::<Interior>()?;
-            next_page.set_prev_page_id(Some(right_page_id));
-        }
-
-        Ok(PendingSplit {
-            separator: propagated_separator,
-            left_page_id: parent_frame.page_id,
-            right_page_id,
-        })
-    }
-
-    /// Installs a new interior root in the existing root page after the old root split.
-    fn install_new_root(&mut self, pending: PendingSplit) -> StorageResult<()> {
-        let root_page_id = self.root_page_id();
-        debug_assert_eq!(
-            pending.left_page_id, root_page_id,
-            "root split must report the old root page as its left page"
-        );
-
-        let root_pin = self.page_cache.fetch_page(root_page_id)?;
-        let root_snapshot = {
-            let root_guard = root_pin.read()?;
-            *root_guard.page()
-        };
-
-        let (left_page_id, left_page_pin) = self.page_cache.new_page()?;
-        {
-            let mut left_guard = left_page_pin.write()?;
-            *left_guard.page_mut() = root_snapshot;
-        }
-        self.relink_copied_root_left_child(left_page_id, pending.right_page_id)?;
-
-        let mut root_guard = root_pin.write()?;
-        let mut root_page = RawInterior::<Write<'_>>::initialize_with_rightmost(
-            root_guard.page_mut(),
-            pending.right_page_id,
-        );
-        self.insert_interior_payload_at(&mut root_page, 0, left_page_id, &pending.separator)?;
-        self.remap_positioned_page(root_page_id, left_page_id);
-        Ok(())
-    }
-
-    fn relink_copied_root_left_child(
-        &self,
-        left_page_id: PageId,
-        right_page_id: PageId,
-    ) -> StorageResult<()> {
-        let left_pin = self.page_cache.fetch_page(left_page_id)?;
-        let mut left_guard = left_pin.write()?;
-        match read_page_kind(left_guard.page(), left_page_id)? {
-            PageKind::RawLeaf => {
-                {
-                    let mut leaf = left_guard.open_mut::<Leaf>()?;
-                    leaf.set_prev_page_id(None);
-                    leaf.set_next_page_id(Some(right_page_id));
-                }
-                self.set_leaf_prev_page_id(right_page_id, Some(left_page_id))?;
-            }
-            PageKind::RawInterior => {
-                {
-                    let mut interior = left_guard.open_mut::<Interior>()?;
-                    interior.set_prev_page_id(None);
-                    interior.set_next_page_id(Some(right_page_id));
-                }
-                self.set_interior_prev_page_id(right_page_id, Some(left_page_id))?;
-            }
-        }
-        Ok(())
-    }
-
-    fn remap_positioned_page(&mut self, from_page_id: PageId, to_page_id: PageId) {
-        if let CursorState::Positioned { page_id, slot_index } = self.state
-            && page_id == from_page_id
-        {
-            self.set_positioned_state(to_page_id, slot_index);
-        }
-    }
-}
-
-/// Allocates and initializes a brand-new empty raw root leaf page.
-pub(crate) fn initialize_empty_root<S: PageStore>(
-    page_cache: &PageCache<S>,
-) -> StorageResult<PageId> {
-    let (page_id, pin) = page_cache.new_page()?;
-    let mut page = pin.write()?;
-    let _ = RawLeaf::<Write<'_>>::initialize(page.page_mut());
-    Ok(page_id)
-}
-
-/// Verifies that `root_page_id` names a raw leaf or raw interior page.
-pub(crate) fn validate_root_page(
-    page_cache: &PageCache<impl PageStore>,
-    root_page_id: PageId,
-) -> StorageResult<()> {
-    let pin = page_cache.fetch_page(root_page_id)?;
-    let page = pin.read()?;
-    match read_page_kind(page.page(), root_page_id)? {
-        PageKind::RawLeaf => {
-            let _ = page.open::<Leaf>()?;
-        }
-        PageKind::RawInterior => {
-            let _ = page.open::<Interior>()?;
-        }
-    }
-    Ok(())
-}
-
-fn read_page_kind(bytes: &[u8; PAGE_SIZE], page_id: PageId) -> StorageResult<PageKind> {
-    let raw_kind = bytes[KIND_OFFSET];
-    PageKind::from_raw(raw_kind).ok_or({
-        StorageError::Corruption(CorruptionError {
-            component: CorruptionComponent::Page,
-            page_id: Some(page_id),
-            kind: CorruptionKind::UnknownPageKind { actual: raw_kind },
-        })
-    })
-}
-
-fn expect_page_kind(
-    bytes: &[u8; PAGE_SIZE],
-    page_id: PageId,
-    expected: PageKind,
-    expected_name: &'static str,
-) -> StorageResult<()> {
-    let raw_kind = bytes[KIND_OFFSET];
-    let actual = read_page_kind(bytes, page_id)?;
-    if actual == expected {
-        Ok(())
-    } else {
-        Err(StorageError::Corruption(CorruptionError {
-            component: CorruptionComponent::Page,
-            page_id: Some(page_id),
-            kind: CorruptionKind::InvalidPageKind { expected: expected_name, actual: raw_kind },
-        }))
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::collections::BTreeMap;
-
-    use fastrand::Rng;
-    use tempfile::NamedTempFile;
-
-    use super::*;
-    use crate::core::disk_manager::DiskManager;
-    use crate::core::error::LimitExceededError;
-    use crate::core::memory_page_store::MemoryPageStore;
-
-    const KEY_LEN_RANGE: std::ops::RangeInclusive<usize> = 8..=192;
-    const VALUE_LEN_RANGE: std::ops::RangeInclusive<usize> = 8..=PAGE_SIZE * 3;
-    const INLINE_VALUE_LEN_RANGE: std::ops::RangeInclusive<usize> = 8..=512;
-    const TARGET_HEIGHT: usize = 4;
-    const MAX_RECORDS: usize = 50_000;
-
-    fn random_bytes(rng: &mut Rng, len: usize) -> Vec<u8> {
-        let mut bytes = vec![0; len];
-        rng.fill(&mut bytes);
-        bytes
-    }
-
-    fn random_unique_cell(
-        rng: &mut Rng,
-        expected: &BTreeMap<Vec<u8>, Vec<u8>>,
-    ) -> (Vec<u8>, Vec<u8>) {
-        loop {
-            let key_len = rng.usize(KEY_LEN_RANGE);
-            let value_len = if rng.u8(0..32) == 0 {
-                rng.usize(VALUE_LEN_RANGE)
-            } else {
-                rng.usize(INLINE_VALUE_LEN_RANGE)
-            };
-            let key = random_bytes(rng, key_len);
-            if expected.contains_key(&key) {
-                continue;
-            }
-
-            let value = random_bytes(rng, value_len);
-            return (key, value);
-        }
-    }
-
-    fn assert_supported_cell(key: &[u8], value: &[u8]) {
-        assert!(
-            key.len() + value.len() <= u16::MAX as usize,
-            "leaf payload should fit the current u16 payload-length field"
-        );
-        assert!(
-            key.len() <= u16::MAX as usize,
-            "interior separator should fit the current u16 payload-length field"
-        );
-    }
-
-    fn memory_tree_cursor(cache_frames: usize) -> TreeCursor<MemoryPageStore> {
-        let page_cache = PageCache::new(MemoryPageStore::new(), cache_frames).unwrap();
-        let root_page_id = initialize_empty_root(&page_cache).unwrap();
-        TreeCursor::new(page_cache, root_page_id)
-    }
-
-    fn tree_height<S: PageStore>(cursor: &TreeCursor<S>) -> StorageResult<usize> {
-        let mut height = 1;
-        let mut page_id = cursor.root_page_id();
-
-        loop {
-            let pin = cursor.page_cache.fetch_page(page_id)?;
-            let next_page_id = {
-                let page = pin.read()?;
-                match read_page_kind(page.page(), page_id)? {
-                    PageKind::RawLeaf => {
-                        let _ = page.open::<Leaf>()?;
-                        return Ok(height);
-                    }
-                    PageKind::RawInterior => {
-                        let interior = page.open::<Interior>()?;
-                        if interior.slot_count() == 0 {
-                            interior.rightmost_child()
-                        } else {
-                            interior.cell(0)?.left_child()?
-                        }
-                    }
-                }
-            };
-
-            height += 1;
-            page_id = next_page_id;
-        }
-    }
-
-    #[test]
-    fn root_page_id_stays_stable_after_root_splits() {
-        let mut cursor = memory_tree_cursor(256);
-        let root_page_id = cursor.root_page_id();
-        let mut expected = BTreeMap::new();
-
-        for index in 0..256_u16 {
-            let key = oversized_key(index);
-            let value = format!("value-{index}").into_bytes();
-            cursor.insert(&key, &value).unwrap();
-            expected.insert(key, value);
-            assert_eq!(cursor.root_page_id(), root_page_id);
-
-            if tree_height(&cursor).unwrap() >= 3 {
-                break;
-            }
-        }
-
-        assert!(tree_height(&cursor).unwrap() >= 3, "test setup should split an interior root");
-        for (key, value) in &expected {
-            let record = cursor.get(key).unwrap().expect("inserted key should be readable");
-            assert_record_matches(&record, key, value);
-        }
-    }
-
-    #[test]
-    fn root_page_id_stays_stable_after_root_shrink() {
-        let mut cursor = memory_tree_cursor(256);
-        let root_page_id = cursor.root_page_id();
-        let mut keys = Vec::new();
-
-        for index in 0..512_u32 {
-            let key = index.to_be_bytes().to_vec();
-            cursor.insert(&key, b"value").unwrap();
-            keys.push(key);
-        }
-        assert!(tree_height(&cursor).unwrap() >= 2, "test setup should split the root");
-
-        for key in keys {
-            cursor.delete(&key).unwrap();
-            assert_eq!(cursor.root_page_id(), root_page_id);
-        }
-
-        assert_eq!(cursor.root_page_id(), root_page_id);
-        assert_eq!(tree_height(&cursor).unwrap(), 1);
-        assert!(!cursor.seek_to_first().unwrap());
-    }
-
-    #[ignore = "slow because of fsync"]
-    #[test]
-    // Builds a four-level raw B+ tree from deterministic random inline cells,
-    // proving leaf splits, repeated interior split propagation, exact-key
-    // lookups, and forward/backward sorted cursor scans.
-    fn random_insert_get_simulation_with_oversized_values_reaches_four_levels() {
-        let mut cursor = memory_tree_cursor(256);
-        let mut rng = Rng::with_seed(0xd47a_ba5e_b7ee_2026);
-        let mut expected = BTreeMap::new();
-        let mut cells = Vec::new();
-        let mut previous_height = tree_height(&cursor).unwrap();
-        let mut saw_leaf_root_split = false;
-        let mut saw_first_interior_root_split = false;
-        let mut saw_repeated_interior_split_propagation = false;
-
-        assert_eq!(previous_height, 1);
-
-        for _ in 0..MAX_RECORDS {
-            let (key, value) = random_unique_cell(&mut rng, &expected);
-            assert_supported_cell(&key, &value);
-
-            cursor.insert(&key, &value).unwrap();
-            assert!(expected.insert(key.clone(), value.clone()).is_none());
-            cells.push((key, value));
-
-            let height = tree_height(&cursor).unwrap();
-            saw_leaf_root_split |= previous_height == 1 && height == 2;
-            saw_first_interior_root_split |= previous_height == 2 && height == 3;
-            saw_repeated_interior_split_propagation |= previous_height == 3 && height == 4;
-            previous_height = height;
-
-            if height == TARGET_HEIGHT {
-                break;
-            }
-        }
-
-        assert!(saw_leaf_root_split, "tree should split the root leaf");
-        assert!(saw_first_interior_root_split, "tree should split an interior root");
-        assert!(
-            saw_repeated_interior_split_propagation,
-            "tree should propagate an interior split through an existing interior level"
-        );
-        assert_eq!(
-            previous_height, TARGET_HEIGHT,
-            "simulation should reach {TARGET_HEIGHT} tree levels within {MAX_RECORDS} inserts"
-        );
-
-        for (key, value) in &cells {
-            assert_eq!(expected.get(key).map(Vec::as_slice), Some(value.as_slice()));
-
-            let record = cursor.get(key).unwrap().expect("inserted tree key should be present");
-            assert_record_matches(&record, key, value);
-        }
-
-        assert_forward_scan_matches(&mut cursor, &expected);
-        assert_reverse_scan_matches(&mut cursor, &expected);
-        assert_eq!(expected.len(), cells.len());
-    }
-
-    #[ignore = "slow because of fsync"]
-    #[test]
-    fn random_insert_delete_simulation_empties_tree_after_random_delete_order() {
-        let mut cursor = memory_tree_cursor(256);
-        let mut rng = Rng::with_seed(0x9dd0_c312_741f_2026);
-        let mut expected = BTreeMap::new();
-
-        const INSERT_COUNT: usize = 200;
-        for _ in 0..INSERT_COUNT {
-            let (key, value) = random_unique_cell(&mut rng, &expected);
-            assert_supported_cell(&key, &value);
-            cursor.insert(&key, &value).unwrap();
-            expected.insert(key, value);
-        }
-        assert!(!expected.is_empty(), "simulation should create at least one record");
-
-        let mut delete_order: Vec<Vec<u8>> = expected.keys().cloned().collect();
-        let sorted_order = delete_order.clone();
-        rng.shuffle(&mut delete_order);
-        while delete_order == sorted_order {
-            rng.shuffle(&mut delete_order);
-        }
-
-        for key in &delete_order {
-            cursor.delete(key).unwrap();
-        }
-
-        assert!(!cursor.seek_to_first().unwrap(), "tree should be empty after deleting all keys");
-        assert!(cursor.current().unwrap().is_none(), "empty tree cursor should have no current");
-        for key in expected.keys() {
-            assert!(cursor.get(key).unwrap().is_none(), "deleted key should not be found");
-        }
-    }
-
-    #[ignore = "slow because of fsync"]
-    #[test]
-    fn random_insert_update_simulation_replaces_values_for_all_keys() {
-        let mut cursor = memory_tree_cursor(256);
-        let mut rng = Rng::with_seed(0xa11d_47e5_2026_0425);
-        let mut expected = BTreeMap::new();
-
-        const INSERT_COUNT: usize = 200;
-        const OVERSIZED_EVERY_NTH: usize = 9;
-
-        for index in 0..INSERT_COUNT {
-            let (key, mut value) = random_unique_cell(&mut rng, &expected);
-            if index % OVERSIZED_EVERY_NTH == 0 {
-                let oversized_len = PAGE_SIZE + rng.usize(1..=PAGE_SIZE);
-                value = random_bytes(&mut rng, oversized_len);
-            }
-
-            assert_supported_cell(&key, &value);
-            cursor.insert(&key, &value).unwrap();
-            expected.insert(key, value);
-        }
-
-        assert!(
-            expected.values().any(|value| value.len() > PAGE_SIZE),
-            "simulation setup should include oversized values"
-        );
-
-        let mut keys: Vec<Vec<u8>> = expected.keys().cloned().collect();
-        rng.shuffle(&mut keys);
-        let mut updated = BTreeMap::new();
-        for (index, key) in keys.iter().enumerate() {
-            let old_value = expected.get(key).unwrap();
-            let new_value = loop {
-                let value_len = if index % OVERSIZED_EVERY_NTH == 0 {
-                    PAGE_SIZE + rng.usize(1..=PAGE_SIZE)
-                } else {
-                    rng.usize(INLINE_VALUE_LEN_RANGE)
-                };
-                let candidate = random_bytes(&mut rng, value_len);
-                if candidate != *old_value {
-                    break candidate;
-                }
-            };
-
-            assert_supported_cell(key, &new_value);
-            cursor.update(key, &new_value).unwrap();
-            updated.insert(key.clone(), new_value);
-        }
-
-        assert_eq!(updated.len(), expected.len());
-        assert!(
-            updated.values().any(|value| value.len() > PAGE_SIZE),
-            "updated values should include oversized entries"
-        );
-
-        for (key, value) in &updated {
-            let record = cursor.get(key).unwrap().expect("updated key should still exist");
-            assert_record_matches(&record, key, value);
-        }
-
-        assert_forward_scan_matches(&mut cursor, &updated);
-        assert_reverse_scan_matches(&mut cursor, &updated);
-    }
-
-    #[ignore = "slow because of fsync"]
-    #[test]
-    fn random_mixed_operation_simulation_matches_btreemap_model() {
-        let mut cursor = memory_tree_cursor(256);
-        let mut rng = Rng::with_seed(0x741e_5afe_2026_0429);
-        let mut expected = BTreeMap::new();
-
-        const STEPS: usize = 3_000;
-        const MIN_LIVE_KEYS: usize = 128;
-        const VERIFY_EVERY_NTH: usize = 50;
-        const OVERSIZED_EVERY_NTH: usize = 31;
-
-        for step in 0..STEPS {
-            let operation = if expected.len() < MIN_LIVE_KEYS { 0 } else { rng.u8(0..100) };
-            match operation {
-                0..=34 => {
-                    let (key, mut value) = random_unique_cell(&mut rng, &expected);
-                    if step % OVERSIZED_EVERY_NTH == 0 {
-                        let oversized_len = PAGE_SIZE + rng.usize(1..=PAGE_SIZE);
-                        value = random_bytes(&mut rng, oversized_len);
-                    }
-
-                    assert_supported_cell(&key, &value);
-                    cursor.insert(&key, &value).unwrap();
-                    assert!(expected.insert(key.clone(), value.clone()).is_none());
-                    let record = cursor.get(&key).unwrap().expect("inserted key should be found");
-                    assert_record_matches(&record, &key, &value);
-                }
-                35..=64 => {
-                    if expected.is_empty() || rng.bool() {
-                        let (key, _) = random_unique_cell(&mut rng, &expected);
-                        assert!(cursor.get(&key).unwrap().is_none());
-                    } else {
-                        let key = random_existing_key(&mut rng, &expected);
-                        let expected_value = expected.get(key).unwrap();
-                        let record = cursor
-                            .get(key)
-                            .unwrap()
-                            .unwrap_or_else(|| panic!("model key should exist at step {step}"));
-                        assert_record_matches(&record, key, expected_value);
-                    }
-                }
-                65..=84 => {
-                    let key = random_existing_key(&mut rng, &expected).to_vec();
-                    let old_value = expected.get(&key).unwrap();
-                    let new_value = loop {
-                        let value_len = if step % OVERSIZED_EVERY_NTH == 0 {
-                            PAGE_SIZE + rng.usize(1..=PAGE_SIZE)
-                        } else {
-                            rng.usize(INLINE_VALUE_LEN_RANGE)
-                        };
-                        let candidate = random_bytes(&mut rng, value_len);
-                        if candidate != *old_value {
-                            break candidate;
-                        }
-                    };
-
-                    assert_supported_cell(&key, &new_value);
-                    cursor.update(&key, &new_value).unwrap();
-                    expected.insert(key.clone(), new_value.clone());
-                    let record = cursor.get(&key).unwrap().expect("updated key should still exist");
-                    assert_record_matches(&record, &key, &new_value);
-                }
-                _ => {
-                    let key = random_existing_key(&mut rng, &expected).to_vec();
-                    cursor.delete(&key).unwrap();
-                    assert!(expected.remove(&key).is_some());
-                    assert!(cursor.get(&key).unwrap().is_none(), "deleted key should not be found");
-                }
-            }
-
-            if step % VERIFY_EVERY_NTH == 0 {
-                assert_tree_gets_match_model(&mut cursor, &expected);
-            }
-        }
-
-        assert_tree_gets_match_model(&mut cursor, &expected);
-    }
-
-    fn oversized_key(index: u16) -> Vec<u8> {
-        let mut key = vec![0; PAGE_SIZE + 256];
-        key[..2].copy_from_slice(&index.to_be_bytes());
-        for byte in &mut key[2..] {
-            *byte = (index % 251) as u8;
-        }
-        key
-    }
-
-    #[test]
-    fn insert_get_supports_oversized_keys_promoted_to_interior_pages() {
-        let mut cursor = memory_tree_cursor(256);
-        let mut expected = BTreeMap::new();
-
-        for index in 0..48 {
-            let key = oversized_key(index);
-            let value = format!("value-{index}").into_bytes();
-            assert_supported_cell(&key, &value);
-            cursor.insert(&key, &value).unwrap();
-            expected.insert(key, value);
-        }
-
-        assert!(tree_height(&cursor).unwrap() >= 2, "large keys should force a root split");
-        for (key, value) in &expected {
-            let record = cursor.get(key).unwrap().expect("inserted oversized key should exist");
-            assert_record_matches(&record, key, value);
-        }
-        assert_forward_scan_matches(&mut cursor, &expected);
-    }
-
-    #[test]
-    fn failed_interior_rewrite_leaves_page_unchanged() {
-        let page_cache = PageCache::new(MemoryPageStore::new(), 16).unwrap();
-        let (page_id, pin) = page_cache.new_page().unwrap();
-        {
-            let mut guard = pin.write().unwrap();
-            let mut interior =
-                RawInterior::<Write<'_>>::initialize_with_rightmost(guard.page_mut(), 2);
-            interior.insert_payload_at(0, 0, b"stable".len(), None, b"stable").unwrap();
-        }
-        drop(pin);
-
-        let cursor = TreeCursor::new(page_cache.clone(), page_id);
-        let original_page = {
-            let pin = page_cache.fetch_page(page_id).unwrap();
-            let page = pin.read().unwrap();
-            *page.page()
-        };
-        let children: Vec<_> = (0..16)
-            .map(|index| ChildEntry {
-                page_id: 100 + index,
-                max_key: Some(vec![index as u8; PAGE_SIZE]),
-            })
-            .collect();
-
-        let result = cursor.rewrite_interior_page(page_id, &children, None, None);
-
-        assert!(matches!(
-            result,
-            Err(StorageError::LimitExceeded(LimitExceededError::PageFull { .. }))
-        ));
-        let rewritten_page = {
-            let pin = page_cache.fetch_page(page_id).unwrap();
-            let page = pin.read().unwrap();
-            *page.page()
-        };
-        assert_eq!(rewritten_page, original_page);
-    }
-
-    #[test]
-    fn unchanged_path_separator_refresh_does_not_grow_file() {
-        let file = NamedTempFile::new().unwrap();
-        let disk_manager = DiskManager::new(file.path()).unwrap();
-        let page_cache = PageCache::new(disk_manager, 256).unwrap();
-        let root_page_id = initialize_empty_root(&page_cache).unwrap();
-        let mut cursor = TreeCursor::new(page_cache, root_page_id);
-        let mut expected = BTreeMap::new();
-
-        for index in 0..96 {
-            let key = oversized_key(index);
-            let value = format!("value-{index}").into_bytes();
-            cursor.insert(&key, &value).unwrap();
-            expected.insert(key, value);
-        }
-
-        let key = expected.keys().next().expect("test setup should create records");
-        let (_, tree_path) = cursor.leaf_page_path_for_key(key).unwrap();
-        assert!(!tree_path.is_empty(), "large keys should force interior separators");
-        cursor.refresh_path_separators(&tree_path).unwrap();
-        let file_len_before = file.path().metadata().unwrap().len();
-
-        cursor.refresh_path_separators(&tree_path).unwrap();
-
-        let file_len_after = file.path().metadata().unwrap().len();
-        assert_eq!(file_len_after, file_len_before);
-    }
-
-    #[test]
-    fn inline_record_is_page_resident() {
-        let mut cursor = memory_tree_cursor(4);
-
-        cursor.insert(b"alpha", b"value").unwrap();
-
-        let record = cursor.get(b"alpha").unwrap().expect("inline record should exist");
-        assert!(matches!(record.storage, RecordStorage::PageResident { .. }));
-        assert_record_matches(&record, b"alpha", b"value");
-    }
-
-    #[test]
-    fn overflow_record_is_materialized() {
-        let mut cursor = memory_tree_cursor(4);
-        let value = vec![42; PAGE_SIZE];
-
-        cursor.insert(b"alpha", &value).unwrap();
-
-        let record = cursor.get(b"alpha").unwrap().expect("overflow record should exist");
-        assert!(matches!(record.storage, RecordStorage::Materialized { .. }));
-        assert_record_matches(&record, b"alpha", &value);
-    }
-
-    #[test]
-    fn inline_record_converts_to_owned_snapshot() {
-        let mut cursor = memory_tree_cursor(4);
-
-        cursor.insert(b"alpha", b"value").unwrap();
-
-        let owned = cursor
-            .get(b"alpha")
-            .unwrap()
-            .expect("inline record should exist")
-            .to_owned_record()
-            .unwrap();
-        assert_owned_record_matches(&owned, b"alpha", b"value");
-    }
-
-    #[test]
-    fn binary_search_supports_inline_key_with_overflow_value() {
-        let mut cursor = memory_tree_cursor(8);
-        let value = vec![7; PAGE_SIZE];
-
-        cursor.insert(b"alpha", b"small").unwrap();
-        cursor.insert(b"bravo", &value).unwrap();
-        cursor.insert(b"charlie", b"small").unwrap();
-
-        let record = cursor.get(b"bravo").unwrap().expect("overflow value key should exist");
-        assert_record_matches(&record, b"bravo", &value);
-        assert!(cursor.get(b"between").unwrap().is_none());
-    }
-
-    #[test]
-    fn binary_search_supports_oversized_key_with_overflow_value() {
-        let mut cursor = memory_tree_cursor(16);
-        let key = oversized_key(7);
-        let value = vec![11; PAGE_SIZE];
-
-        cursor.insert(&key, &value).unwrap();
-
-        let record = cursor.get(&key).unwrap().expect("oversized key should exist");
-        assert_record_matches(&record, &key, &value);
-    }
-
-    #[test]
-    fn binary_search_supports_oversized_interior_separator_keys() {
-        let mut cursor = memory_tree_cursor(256);
-        let mut expected = BTreeMap::new();
-
-        for index in 0..48 {
-            let key = oversized_key(index);
-            let value = format!("value-{index}").into_bytes();
-            cursor.insert(&key, &value).unwrap();
-            expected.insert(key, value);
-        }
-
-        assert!(tree_height(&cursor).unwrap() >= 2, "large keys should force interior routing");
-        for (key, value) in &expected {
-            assert!(cursor.seek_to_key(key).unwrap(), "seek_to_key should find oversized key");
-            let record = cursor.current().unwrap().expect("seek_to_key should position cursor");
-            assert_record_matches(&record, key, value);
-        }
-    }
-
-    fn assert_record_matches<S: PageStore>(
-        record: &Record<S>,
-        expected_key: &[u8],
-        expected_value: &[u8],
-    ) {
-        record
-            .with_key_value(|actual_key, actual_value| {
-                assert_eq!(actual_key, expected_key);
-                assert_eq!(actual_value, expected_value);
-            })
-            .unwrap();
-    }
-
-    fn assert_owned_record_matches(
-        record: &OwnedRecord,
-        expected_key: &[u8],
-        expected_value: &[u8],
-    ) {
-        record.with_key_value(|actual_key, actual_value| {
-            assert_eq!(actual_key, expected_key);
-            assert_eq!(actual_value, expected_value);
-        });
-    }
-
-    fn random_existing_key<'a>(
-        rng: &mut Rng,
-        expected: &'a BTreeMap<Vec<u8>, Vec<u8>>,
-    ) -> &'a Vec<u8> {
-        let index = rng.usize(0..expected.len());
-        expected.keys().nth(index).expect("model should contain keys")
-    }
-
-    fn assert_tree_gets_match_model<S: PageStore>(
-        cursor: &mut TreeCursor<S>,
-        expected: &BTreeMap<Vec<u8>, Vec<u8>>,
-    ) {
-        if expected.is_empty() {
-            assert!(!cursor.seek_to_first().unwrap(), "empty model should match empty tree");
-            assert!(
-                cursor.current().unwrap().is_none(),
-                "empty tree cursor should have no current"
-            );
-            return;
-        }
-
-        for (key, value) in expected {
-            let record = cursor.get(key).unwrap().expect("model key should exist in tree");
-            assert_record_matches(&record, key, value);
-        }
-    }
-
-    fn assert_forward_scan_matches<S: PageStore>(
-        cursor: &mut TreeCursor<S>,
-        expected: &BTreeMap<Vec<u8>, Vec<u8>>,
-    ) {
-        assert!(cursor.seek_to_first().unwrap(), "tree should not be empty");
-
-        let mut expected_entries = expected.iter();
-        let (first_key, first_value) = expected_entries.next().unwrap();
-        let first_record = cursor.current().unwrap().expect("seek_to_first should position cursor");
-        assert_record_matches(&first_record, first_key, first_value);
-
-        let mut scanned = 1;
-        for (key, value) in expected_entries {
-            let record = cursor.next_record().unwrap().expect("forward scan ended early");
-            assert_record_matches(&record, key, value);
-            scanned += 1;
-        }
-
-        assert!(cursor.next_record().unwrap().is_none());
-        assert_eq!(scanned, expected.len());
-    }
-
-    fn assert_reverse_scan_matches<S: PageStore>(
-        cursor: &mut TreeCursor<S>,
-        expected: &BTreeMap<Vec<u8>, Vec<u8>>,
-    ) {
-        let (last_key, _) = expected.iter().next_back().unwrap();
-        let last_record = cursor.get(last_key).unwrap().expect("last key should be present");
-
-        let mut expected_entries = expected.iter().rev();
-        let (key, value) = expected_entries.next().unwrap();
-        assert_record_matches(&last_record, key, value);
-
-        let mut scanned = 1;
-        for (key, value) in expected_entries {
-            let record = cursor.prev_record().unwrap().expect("reverse scan ended early");
-            assert_record_matches(&record, key, value);
-            scanned += 1;
-        }
-
-        assert!(cursor.prev_record().unwrap().is_none());
-        assert_eq!(scanned, expected.len());
-    }
 }
